% Generated by roxygen2 (4.0.2): do not edit by hand
\name{affine_matrix}
\alias{affine_matrix}
\title{Construct an affine transformation matrix from user input}
\usage{
affine_matrix()
}
\value{
A 4x4 numeric matrix, produced by
\code{\link{cmtk.dof2mat}}.
}
\description{
Prompts the user for translation, rotation, scaling,
shearing and centre details of transformation and converts
these into an affine transformation matrix, compatible with
CMTK. Parameters should be seperated by single spaces.
}

