% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.landmarks.amira}
\alias{read.landmarks.amira}
\alias{write.landmarks.amira}
\title{Read and Write Amira Landmarks files}
\usage{
read.landmarks.amira(file, CoordinatesOnly = TRUE, Verbose = FALSE)

write.landmarks.amira(x, file)
}
\arguments{
  \item{file}{The file to read or write}

  \item{CoordinatesOnly}{Only return Coordinates of points}

  \item{Verbose}{Whether to write status messages}

  \item{x}{Nx3 matrix or dataframe of landmark coordinates
  or a list with two matrices, one for landmarks in each
  space.}
}
\value{
for read.amiralandmarks a matrix for an unpaired landmark
set or a list of length 2 for a paired landmark set.
}
\description{
Read and Write Amira Landmarks files
}
\seealso{
Other amira: \code{\link{amiratype}}; \code{\link{hxsurf}},
\code{\link{read.hxsurf}}; \code{\link{is.amiramesh}};
\code{\link{read.amiramesh}},
\code{\link{read.amiramesh.header}};
\code{\link{write.hxsurf}}
}

