% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amiramesh-io.R
\name{amiratype}
\alias{amiratype}
\title{Return the type of an AmiraMesh file on disk or a parsed header}
\usage{
amiratype(x, bytes = NULL)
}
\arguments{
\item{x}{Path to files on disk or a single pre-parsed parameter list}

\item{bytes}{A raw vector containing at least 11 bytes from the start of the
file.}
}
\value{
character vector (NA_character_ when file invalid)
}
\description{
Return the type of an AmiraMesh file on disk or a parsed header
}
\details{
Note that when checking a file we first test if it is an AmiraMesh 
  file (fast, especially when \code{bytes!=NULL}) before reading the header 
  and determining content type (slow).
}
\seealso{
Other amira:
\code{\link[=is.amiramesh]{is.amiramesh()}},
\code{\link[=read.amiramesh]{read.amiramesh()}},
\code{\link[=read.hxsurf]{read.hxsurf()}},
\code{\link[=write.hxsurf]{write.hxsurf()}}
}
\concept{amira}
