% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/neuronlistfh.R
\name{as.neuronlist.neuronlistfh}
\alias{as.neuronlist.neuronlistfh}
\title{convert neuronlistfh to a regular (in memory) neuronlist}
\usage{
\method{as.neuronlist}{neuronlistfh}(l, ...)
}
\arguments{
  \item{l}{An existing list or a single neuron to start a
  list}

  \item{...}{Additional arguments passed to methods}
}
\description{
convert neuronlistfh to a regular (in memory) neuronlist
}

