\name{as.neuronlistfh.neuronlist}
\alias{as.neuronlistfh.neuronlist}
\title{convert a regular neuronlist to one backed by a filehash object}
\usage{
\method{as.neuronlistfh}{neuronlist}(l, df, ..., dbName = "nldb",
  type = "RDS")
}
\description{
convert a regular neuronlist to one backed by a filehash
object
}

