% Generated by roxygen2 (4.0.1): do not edit by hand
\name{c.neuronlist}
\alias{c.neuronlist}
\title{Combine multiple neuronlists into a single list}
\usage{
\method{c}{neuronlist}(..., columns = c("all", "matching"),
  recursive = FALSE)
}
\arguments{
  \item{...}{neuronlists to combine}

  \item{columns}{Either \code{'all'}, the default, or
  \code{'matching'}. See details.}

  \item{recursive}{Presently ignored}
}
\description{
Combine multiple neuronlists into a single list
}
\details{
Uses \code{\link[plyr]{rbind.fill}} to join any attached
dataframes.  When \code{columns='all'}, the default,
missing values are replaced with NAs. When
\code{columns='matching'}, only columns present in all
attached data.frames are kept.
}
\examples{
stopifnot(all.equal(kcs20[1:2],c(kcs20[1],kcs20[2])))
}
\seealso{
\code{\link[base]{c}}
}

