% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat-data.R
\docType{data}
\name{Cell07PNs}
\alias{Cell07PNs}
\title{Cell07PNs: 40 Sample Projection Neurons from Jefferis, Potter et al 2007}
\description{
These R lists (which have additional class neuronlist) contain 40 traced
olfactory projection neurons from Jefferis, Potter et al 2007 that have been
transformed onto the IS2 template brain (Cachero, Ostrovsky et al 2010).
}
\examples{
head(Cell07PNs)
table(with(Cell07PNs,Glomerulus))
}
\references{
Jefferis G.S.X.E., Potter C.J., Chan A.M., Marin E.C., Rohlfing 
T., Maurer C.R.J., and Luo L. (2007). Comprehensive maps of Drosophila higher
olfactory centers: spatially segregated fruit and pheromone representation. 
Cell 128 (6), 1187--1203. 
\doi{10.1016/j.cell.2007.01.040}

Cachero S., Ostrovsky A.D., Yu J.Y., Dickson B.J., and Jefferis G.S.X.E. 
(2010). Sexual dimorphism in the fly brain. Curr Biol 20 (18), 1589--601. 
\doi{10.1016/j.cub.2010.07.045}
}
\seealso{
\code{\link{head.neuronlist}}, \code{\link{with.neuronlist}},
  \code{\link{dl1neuron}}

Other nat-data:
\code{\link{MBL.surf}},
\code{\link{kcs20}}
}
\concept{nat-data}
