% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmtk_io.R
\name{cmtk.extract_affine}
\alias{cmtk.extract_affine}
\title{Extract affine registration from CMTK registration file or in-memory list}
\usage{
cmtk.extract_affine(r, outdir)
}
\arguments{
\item{r}{A registration list or path to file on disk}

\item{outdir}{Optional path to output file}
}
\value{
When \code{outdir} is missing a list containing the registration
  parameters Otherwise \code{NULL} invisibly.
}
\description{
Extract affine registration from CMTK registration file or in-memory list
}
\seealso{
\code{\link{cmtkreglist}}

Other cmtk-io:
\code{\link[=read.cmtk]{read.cmtk()}},
\code{\link[=read.cmtkreg]{read.cmtkreg()}},
\code{\link[=write.cmtk]{write.cmtk()}},
\code{\link[=write.cmtkreg]{write.cmtkreg()}}
}
\concept{cmtk-io}
