% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cmtk.targetvolume}
\alias{cmtk.targetvolume}
\alias{cmtk.targetvolume.default}
\alias{cmtk.targetvolume.im3d}
\title{Defines a target volume for a CMTK reformatx operation}
\usage{
cmtk.targetvolume(target, ...)

\method{cmtk.targetvolume}{im3d}(target, ...)

\method{cmtk.targetvolume}{default}(target, ...)
}
\arguments{
  \item{target}{A character vector specifying a file, an
  \code{im3d} object or a 6-or 9-vector defining a grid in
  the form Nx,Ny,Nz,dX,dY,dZ,[Ox,Oy,Oz].}

  \item{...}{additional arguments passed to methods}
}
\value{
a character vector specifying the full cmtk reformatx
'--target' or '--target-grid' argument
}
\description{
Defines a target volume for a CMTK reformatx operation
}
\details{
if the character vector specifies an amiramesh file, it
will be converted to a bare \code{im3d} object and then to
an appropriate '--target-grid' specification.
}

