% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmtk.R
\name{cmtk.version}
\alias{cmtk.version}
\title{Return cmtk version or test for presence of at least a specific version}
\usage{
cmtk.version(minimum = NULL)
}
\arguments{
\item{minimum}{If specified checks that the cmtk version}
}
\value{
returns \code{numeric_version} representation of CMTK version or if 
  minimum is not NULL, returns a logical indicating whether the installed 
  version exceeds the current version. If CMTK is not installed returns NA.
}
\description{
Return cmtk version or test for presence of at least a specific version
}
\details{
NB this function has the side effect of setting an option 
  nat.cmtk.version the first time that it is run in the current R session.
}
\examples{
\dontrun{
cmtk.version()
cmtk.version('3.2.2')
}
}
\seealso{
\code{\link{cmtk.bindir}, \link{cmtk.dof2mat}}
}
