% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.landmarks.cmtk}
\alias{cmtklandmarks}
\alias{read.landmarks.cmtk}
\alias{write.landmarks.cmtk}
\title{Read and Write CMTK landmarks}
\usage{
read.landmarks.cmtk(con)

cmtklandmarks(xyzs)

write.landmarks.cmtk(xyzs, filename, Force = FALSE)
}
\arguments{
  \item{con}{Character vector specifying path or a
  connection (passed straight to \code{read.cmtk})}

  \item{xyzs}{Nx3 matrix of landmarks}

  \item{filename}{Path to write out cmtklandmarks}

  \item{Force}{Whether to overwrite an existing landmarks
  file (default FALSE)}
}
\description{
Read and Write CMTK landmarks

\code{cmtklandmarks} generates in memory list
representation of cmtk landmarks
}
\details{
CMTK landmarks are always unpaired i.e. only contain
information for one brain.
}
\seealso{
Other cmtk-io: \code{\link{read.cmtkreg}};
\code{\link{read.cmtk}}; \code{\link{write.cmtkreg}};
\code{\link{write.cmtk}}
}

