% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{correct_root}
\alias{correct_root}
\title{Interactively re-root neurons (usually to their soma)}
\usage{
correct_root(someneuronlist, brain = NULL)
}
\arguments{
\item{someneuronlist}{a neuron/neuronlist object}

\item{brain}{for context, plot some other objects (e.g. a brain from the
nat.templatebrains package such as FCWB, or any object that may be plotted
using plot3d)}
}
\value{
a matrix of 3D points
}
\description{
Cycle through and manually re-root neurons using an rgl window.
  This will typically be used for manual identification of the soma of a
  neuron.
}
\examples{
\dontrun{
## NB these neurons actually have their somata chopped off
correctedsomas = correct_root(Cell07PNs[1:3])
plot3d(correctedsomas, soma=TRUE)
}
}
