% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{correct_soma}
\alias{correct_soma}
\title{Interactively re-root neurons to their soma}
\usage{
correct_soma(someneuronlist, brain = NULL)
}
\arguments{
\item{someneuronlist}{a neuron/neuronlist object}

\item{brain}{for context, plot some other objects (e.g. a brain from the nat.templatebrains package such as FCWB,
or any object that may be plotted using plot3d)}
}
\value{
a matrix of 3D points
}
\description{
Cycle through and manually re-root neurons to their soma 
using an rgl window
}
\examples{
\dontrun{ 
## Admittedly, these neurns have their somata chopped off! 
correctedsoma = correct_soma(Cell07PNs)
} 
}
