% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat-data.R
\docType{data}
\name{dl1neuron}
\alias{dl1neuron}
\title{Olfactory Projection Neuron reconstructed from EM data}
\format{A \code{neuron} object with additional class \code{catmaidneuron}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
dl1neuron
}
\description{
A DL1 olfactory projection neuuronobject traced in CATMAID from the FAFB
whole brain EM volume. This has a complex morphology that makes a good test
for pruning and simplifcation strategies.
}
\seealso{
\code{\link{Cell07PNs}}
}
\keyword{datasets}
