% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngraph.R
\name{.verify_input_neuron}
\alias{.verify_input_neuron}
\title{Checks whether the input is a neuron}
\usage{
.verify_input_neuron(n)
}
\arguments{
\item{n}{an object}
}
\description{
Checks whether the input is a neuron
}
