% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotprops.R
\name{Ops.dotprops}
\alias{Ops.dotprops}
\title{Arithmetic for dotprops objects}
\usage{
\method{Ops}{dotprops}(e1, e2 = NULL)
}
\arguments{
\item{e1}{A dotprops object}

\item{e2}{A scalar or 3-vector that will be applied to the dotprops object}
}
\value{
A new dotprops object
}
\description{
Arithmetic for dotprops objects
}
\seealso{
\code{\link{scale.dotprops}}, \code{\link{Ops.neuron}}
}
