% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotprops.R
\name{get_topo_features}
\alias{get_topo_features}
\alias{get_distance_to_soma}
\title{Get topological features per each node}
\usage{
get_topo_features(n)

get_distance_to_soma(n)
}
\arguments{
\item{n}{neuron object with soma}
}
\value{
list with distance and Reversed Strahler order features per node.

vector with distances from soma
}
\description{
Assigns to each node a distance from cell body.
}
\examples{
get_topo_features(Cell07PNs[[1]])
get_distance_to_soma(Cell07PNs[[1]])
}
\seealso{
\code{\link{dotprops}}, \code{\link{ngraph}}
}
