% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find.soma}
\alias{find.soma}
\title{Find names of neurons with soma within a 3d selection box (usually drawn in rgl window)}
\usage{
find.soma(sel3dfun = select3d(), indices = names(db),
  db = getOption("nat.default.neuronlist"), threshold = 0)
}
\arguments{
  \item{sel3dfun}{A \code{\link{select3d}} style function
  to indicate if points are within region}

  \item{indices}{Names of neurons to search (defaults to
  all neurons in list)}

  \item{db}{\code{neuronlist} to search. Can also be a
  character vector naming the neuronlist. Defaults to
  \code{options('nat.default.neuronlist')}.}

  \item{threshold}{More than this many points must be
  present in region}
}
\value{
Character vector of names of selected neurons
}
\description{
Find names of neurons with soma within a 3d selection box
(usually drawn in rgl window)
}
\details{
Uses \code{\link{subset.neuronlist}}, so can work on
neurons or neuron lists.
}
\seealso{
\code{\link{select3d}, \link{subset.neuronlist},
\link{find.neuron}}
}

