% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glue_swc}
\alias{glue_swc}
\title{Glue two SWC neurons together}
\usage{
glue_swc(file1, file2, col1 = "red", col2 = "blue", filepath = NULL)
}
\arguments{
  \item{file1}{path to first SWC file.}

  \item{file2}{path to second SWC file.}

  \item{col1}{colour to plot first neuron in.}

  \item{col2}{colour to plot second neuron in.}

  \item{filepath}{if not \code{NULL}, the glued-together
  neuron will be saved in SWC format to this file path.}
}
\value{
A \code{\link{neuron}} object representing the
glued-together SWC files.
}
\description{
Glue two SWC neurons together
}
\details{
Some tracing software / users save each branch of a neuron
as a separate SWC file, which is incorrect and severely
limits the functionality of nat when used with these files.
This function provides a way to 'glue' these neuron
branches back together into one SWC file, via a
user-selection of a point from each neuron that should be
joined together. This function assumes that the SWC files
are completely unbranched, and will give erroneous results
if used on SWC files that contain branches (unless the
branch points are all lower down the tree than the selected
glue point).
}

