% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hxsurf.R
\name{c.hxsurf}
\alias{c.hxsurf}
\title{Concatenate HyperSurface objects}
\usage{
\method{c}{hxsurf}(...)
}
\arguments{
\item{...}{multiple hxsurf objects}
}
\value{
new hxsurf
}
\description{
Concatenate HyperSurface objects
}
\examples{
h1 = as.hxsurf(icosahedron3d(), 'a')
h2 = as.hxsurf(tetrahedron3d()+1, 'b')
h3 = as.hxsurf(icosahedron3d()+3, 'c')
hc = c(h1, h2, h3)
}
