\name{image.im3d}
\alias{image.im3d}
\title{Method to plot spatially calibrated image arrays}
\usage{
\method{image}{im3d}(x, xlim = NULL, ylim = NULL, zlim = NULL,
  plotdims = NULL, flipdims = "y", filled.contour = FALSE, asp = NA,
  axes = FALSE, xlab = NULL, ylab = NULL, nlevels = 20,
  levels = pretty(zlim, nlevels + 1),
  color.palette = colorRampPalette(c("navy", "cyan", "yellow", "red")),
  col = color.palette(length(levels) - 1), ...)
}
\arguments{
  \item{x}{The im3d object containing the data to be
  plotted (\code{NA}s are allowed).}

  \item{zlim}{the minimum and maximum \code{z} values for
  which colors should be plotted, defaulting to the range
  of the finite values of \code{z}. Each of the given
  colors will be used to color an equispaced interval of
  this range. The \emph{midpoints} of the intervals cover
  the range, so that values just outside the range will be
  plotted.}

  \item{xlim,ylim}{ranges for the plotted \code{x} and
  \code{y} values, defaulting to the BoundingBox of x.}

  \item{xlab,ylab}{each a character string giving the
  labels for the x and y axis.  Default to the \sQuote{call
  names} of \code{x} or \code{y}, or to \code{""} if these
  were unspecified.}

  \item{plotdims}{Which dimensions of 3d \code{im3d} object
  to plot (character vector). Defaults to
  \code{c('x','y')}}

  \item{flipdims}{Which dimensions to flip (character
  vector). Defaults to flipping y.}

  \item{filled.contour}{Whether to use a
  \code{\link{filled.contour}} plot instead of a regular
  \code{\link{image}} plot.}

  \item{asp}{Whether to have a a square aspect ratio
  (logical, default: \code{FALSE})}

  \item{nlevels}{The number of colour levels in z}

  \item{levels}{The levels at which to break z values}

  \item{axes}{Whether to plot axes (default: \code{FALSE})}

  \item{col}{a list of colors such as that generated by
  \code{\link{rainbow}}, \code{\link{heat.colors}},
  \code{\link{topo.colors}}, \code{\link{terrain.colors}}
  or similar functions.}

  \item{color.palette}{The colour palette from which
  \code{col} will be selected.}

  \item{\dots}{graphical parameters for \code{\link{plot}}
  or \code{\link{image}} may also be passed as arguments to
  this function. See \code{\link{image.default}}.}
}
\description{
Method to plot spatially calibrated image arrays
}
\examples{
LHMask=read.im3d(system.file('testdata/nrrd/LHMask.nrrd',package='nat'))
image(LHMask[,,10], asp=TRUE)
# useRaster is appreciably quicker in most cases
image(LHMask[,,10], asp=TRUE, useRaster=TRUE)
}

