% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im3d.R
\name{imexpand.grid}
\alias{imexpand.grid}
\title{Convert locations of im3d voxel grid into XYZ coordinates}
\usage{
imexpand.grid(d)
}
\arguments{
\item{d}{An \code{im3d} object}
}
\value{
Nx3 matrix of image coordinates
}
\description{
Convert locations of im3d voxel grid into XYZ coordinates
}
\examples{
d=im3d(,dim=c(2,3,2),origin=c(10,20,30),voxdims=c(1,2,3))
imexpand.grid(d)
}
\seealso{
expand.grid

Other im3d:
\code{\link[=as.im3d]{as.im3d()}},
\code{\link[=boundingbox]{boundingbox()}},
\code{\link[=im3d]{im3d()}},
\code{\link{im3d-coords}},
\code{\link{im3d-io}},
\code{\link[=imslice]{imslice()}},
\code{\link[=is.im3d]{is.im3d()}},
\code{\link[=mask]{mask()}},
\code{\link[=origin]{origin()}},
\code{\link[=projection]{projection()}},
\code{\link[=threshold]{threshold()}},
\code{\link[=unmask]{unmask()}},
\code{\link[=voxdims]{voxdims()}}
}
\concept{im3d}
