% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im3d.R
\name{imslice}
\alias{imslice}
\title{Slice out a 3D subarray (or 2d matrix) from a 3D image array}
\usage{
imslice(x, slice, slicedim = "z", drop = TRUE)
}
\arguments{
\item{x}{An im3d object}

\item{slice}{Indices defining the slices to keep}

\item{slicedim}{Character vector or integer defining axis from which slices 
will be removed.}

\item{drop}{Whether singleton dimensions will be dropped (default: TRUE) 
converting 3D array to 2d matrix.}
}
\description{
Slice out a 3D subarray (or 2d matrix) from a 3D image array
}
\details{
Note the sample locations stored in the x,y,z attributes will be 
  updated appropriately. FIXME: Should we also update bounding box?
}
\seealso{
Other im3d:
\code{\link[=as.im3d]{as.im3d()}},
\code{\link[=boundingbox]{boundingbox()}},
\code{\link[=im3d]{im3d()}},
\code{\link{im3d-coords}},
\code{\link{im3d-io}},
\code{\link[=imexpand.grid]{imexpand.grid()}},
\code{\link[=is.im3d]{is.im3d()}},
\code{\link[=mask]{mask()}},
\code{\link[=origin]{origin()}},
\code{\link[=projection]{projection()}},
\code{\link[=threshold]{threshold()}},
\code{\link[=unmask]{unmask()}},
\code{\link[=voxdims]{voxdims()}}
}
\concept{im3d}
