% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron-io-fiji.R
\name{is.fijitraces}
\alias{is.fijitraces}
\title{Check whether a file is in Fiji's simple neurite tracer format}
\usage{
is.fijitraces(f, bytes = NULL)
}
\arguments{
\item{f}{path to a file on disk}

\item{bytes}{optional raw vector of bytes used for prechecks}
}
\description{
This will check a file on disk to see if it is in Fiji's simple neurite
tracer XML format.
}
\details{
Some prechecks (optionally taking place on a supplied raw vector of bytes)
should weed out nearly all true negatives and identify many true positives
without having to read/parse the file header.
}
