% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im3d.R
\name{is.im3d}
\alias{is.im3d}
\title{Test if an object is of class im3d}
\usage{
is.im3d(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
logical
}
\description{
Test if an object is of class im3d
}
\seealso{
Other im3d:
\code{\link[=as.im3d]{as.im3d()}},
\code{\link[=boundingbox]{boundingbox()}},
\code{\link[=im3d]{im3d()}},
\code{\link{im3d-coords}},
\code{\link{im3d-io}},
\code{\link[=imexpand.grid]{imexpand.grid()}},
\code{\link[=imslice]{imslice()}},
\code{\link[=mask]{mask()}},
\code{\link[=origin]{origin()}},
\code{\link[=projection]{projection()}},
\code{\link[=threshold]{threshold()}},
\code{\link[=unmask]{unmask()}},
\code{\link[=voxdims]{voxdims()}}
}
\concept{im3d}
