% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/neuron-io-neuroml.R
\name{is.neuroml}
\alias{is.neuroml}
\title{Check whether a file is in NeuroML format}
\usage{
is.neuroml(f, bytes = NULL)
}
\arguments{
  \item{f}{path to a file on disk}

  \item{bytes}{optional raw vector of bytes used for
  prechecks}
}
\description{
This will check a file on disk to see if it is in NeuroML
format. Some prechecks (optionally taking place on a
supplied raw vector of bytes) should weed out nearly all
true negatives and identify many true positives without
having to read/parse the file header.
}

