% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrrd-io.R
\name{is.nrrd}
\alias{is.nrrd}
\title{Check if a file is a NRRD file}
\usage{
is.nrrd(f = NULL, bytes = NULL, ReturnVersion = FALSE, TrustSuffix = FALSE)
}
\arguments{
\item{f}{A character vector specifying the path or a raw vector with at least
8 bytes.}

\item{bytes}{optional raw vector of at least 8 bytes from the start of a
single file (used in preference to reading file \code{f}).}

\item{ReturnVersion}{Whether to return the version of the nrrd format in 
which the file is encoded (1-5).}

\item{TrustSuffix}{Whether to trust that a file ending in .nrrd or .nhdr is a
NRRD}
}
\description{
Check if a file is a NRRD file
}
\details{
Note that multiple files can be checked when a character vector of 
  length > 1 is provided, but only one file can be checked when a raw byte 
  array is provided.
}
\seealso{
Other nrrd:
\code{\link[=nrrd.voxdims]{nrrd.voxdims()}},
\code{\link[=read.nrrd]{read.nrrd()}},
\code{\link[=write.nrrd]{write.nrrd()}}
}
\concept{nrrd}
