% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron-io.R
\name{is.swc}
\alias{is.swc}
\title{Test if a file is an SWC format neuron}
\usage{
is.swc(f, TrustSuffix = TRUE)
}
\arguments{
\item{f}{Path to one or more files}

\item{TrustSuffix}{Whether to trust that a file ending in .nrrd or .nhdr is a
NRRD}
}
\value{
logical value
}
\description{
Test if a file is an SWC format neuron
}
\details{
Note that this test is somewhat expensive compared with the other 
  file tests since SWC files do not have a consistent magic value. It
  therefore often has to read and parse the first few lines of the file in 
  order to determine whether they are consistent with the SWC format.
}
\seealso{
\code{\link{read.neuron}}
}
