\name{lapply.neuronlistfh}
\alias{lapply.neuronlistfh}
\title{Apply a function over a neuronlistfh}
\usage{
\method{lapply}{neuronlistfh}(X, FUN, ...)
}
\arguments{
  \item{X}{A neuronlistfh object}

  \item{FUN}{a function to apply to each element of X}

  \item{...}{Arguments for fun passed on to
  lapply.filehash}
}
\description{
Apply a function over a neuronlistfh
}

