% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im3d.R
\name{makeboundingbox}
\alias{makeboundingbox}
\title{Construct a 3D bounding box object}
\usage{
makeboundingbox(x, dims, input = c("boundingbox", "bounds"))
}
\arguments{
\item{x}{A vector or matrix specifying a bounding box}

\item{dims}{The number of voxels in each dimension when x is a BoundingBox 
matrix.}

\item{input}{Whether \code{x} defines the boundingbox or bounds of the image
(see details).}
}
\description{
\code{makeboundingbox} explicitly constructs a 3D bounding box from a set of
6 numbers defining the opposite corners of a cube. Most of the time as an end
user you will want to compute/get the bounding box of objects/vertices using
\code{\link{boundingbox}}.
}
\seealso{
\code{link{boundingbox}}
}
