% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmtk_io.R
\name{trim}
\alias{trim}
\title{nat package internal functions}
\usage{
trim(t)
}
\arguments{
\item{t}{Character vector to trim}
}
\description{
Utility function to trim whitespace at start and end of lines
}

