% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron-plot.R
\name{nclear3d}
\alias{nclear3d}
\title{Clear the rgl or plotly 3D scene}
\usage{
nclear3d(..., plotengine = getOption("nat.plotengine"))
}
\arguments{
\item{...}{Additional arguments passed to
\code{\link[rgl:clear3d]{rgl::clear3d}}}

\item{plotengine}{the plotting backend engine to use either 'rgl' or 'plotly'.}
}
\description{
Clear the rgl or plotly 3D scene
}
\details{
\code{\link{rgl}} and \code{\link[plotly:plot_ly]{plotly}} have
  quite different models for how to handle the active plot. \code{nclear3d}
  and \code{\link{nopen3d}} allow you to treat them more similarly. Use them
  wherever you use the rgl \code{clear3d} and \code{open3d} commands and your
  could she be able to run with both \bold{plotly} or \bold{rgl} as the
  \code{plotengine}.
}
\examples{

\donttest{
nclear3d()
plot3d(Cell07PNs[[1]])
}
}
\seealso{
\code{\link[rgl:clear3d]{rgl::clear3d}}, \code{\link[nat]{plot3d}},
  \code{\link[nat]{plot3d.neuronlist}}, \code{\link[nat]{nopen3d}}
}
