\name{ndigest}
\alias{ndigest}
\alias{ndigest.neuronlistfh}
\title{Calculated normalised digest value for an object}
\usage{
ndigest(x, ...)

\method{ndigest}{neuronlistfh}(x, ...)
}
\description{
The \emph{normalised} digest should exclude any fields or
attributes irrelevant to the core contents of the object
(e.g. timestamps, absolute location of the input files )
}
\details{
ndigest.neuronlistfh
}

