% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron.R
\name{Ops.neuron}
\alias{Ops.neuron}
\title{Arithmetic for neuron coordinates}
\usage{
\method{Ops}{neuron}(e1, e2 = NULL)
}
\arguments{
\item{e1}{a neuron}

\item{e2}{(a numeric vector to multiply neuron coords in neuron)}
}
\value{
modified neuron
}
\description{
If x is a 1-vector or a 3-vector, operate on xyz only
If x is a 4-vector, apply operation to xyz and diameter
}
\examples{
n1<-Cell07PNs[[1]]*2
n2<-Cell07PNs[[1]]*c(2,2,2,1)
stopifnot(all.equal(n1,n2))
n3<-Cell07PNs[[1]]*c(2,2,4)
}
\seealso{
neuron
}
