\docType{methods}
\name{lapply,neuronlistfh-method}
\alias{lapply,neuronlistfh-method}
\title{Methods for neuronlistfh objects}
\usage{
\S4method{lapply}{neuronlistfh}(X, FUN, ...)
}
\arguments{
  \item{X}{A neuronlistfh object}

  \item{FUN}{a function to apply to each element of X}

  \item{...}{Additional arguments for FUN}
}
\description{
Methods for neuronlistfh objects
}

