% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron-plot.R
\name{nopen3d}
\alias{nopen3d}
\title{Open customised rgl window}
\usage{
nopen3d(bgcol = "white", FOV = 0, ...)
}
\arguments{
\item{bgcol}{background colour}

\item{FOV}{field of view}

\item{...}{additional options passed to open3d}
}
\value{
current rgl device
}
\description{
Open customised rgl window
}
\details{
Pan with right button (Ctrl+click), zoom with middle
  (Alt/Meta+click) button. On a Mac trackpad, pan with two fingers
  left-right, zoom with two fingers in-out. Defaults to a white background
  and orthogonal projection (FOV=0)

  Note that sometimes (parts of) objects seem to disappear after panning and
  zooming. See help for \code{\link{pan3d}}.

  \code{\link[rgl]{rgl}} and \code{\link[plotly:plot_ly]{plotly}} have quite
  different models for how to handle the active plot. \code{nopen3d} and
  \code{\link{nclear3d}} allow you to treat them more similarly. Use them
  wherever you use the rgl \code{\link[rgl]{clear3d}} and
  \code{\link[rgl]{open3d}} commands and your
  code should be able to run with both \bold{plotly} or \bold{rgl} as the
  \code{plotengine}.
}
\seealso{
\code{\link[rgl]{open3d}, \link{pan3d}}, \code{\link{nclear3d}}
}
