% Generated by roxygen2 (4.0.1): do not edit by hand
\name{npop3d}
\alias{npop3d}
\title{Remove plotted neurons or other 3d objects}
\usage{
npop3d(x, slow = FALSE, type = "shapes")
}
\arguments{
  \item{x}{rgl ids of objects to remove}

  \item{slow}{Whether to remove neurons one by one (slowly)
  default: FALSE}

  \item{type}{Type of objects to remove see \code{pop3d}.}
}
\description{
The normal usage will not specify \code{x} in which case
the last neurons plotted by \code{plot3d.neuronlist} or any
of its friends will be removed.
}
\seealso{
\code{\link[rgl]{pop3d}}, \code{\link{plot3d.neuronlist}}
}

