% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{nvertices}
\alias{nvertices}
\alias{nvertices.default}
\alias{nvertices.neuronlist}
\title{Find the number of vertices in an object (or each element of a neuronlist)}
\usage{
nvertices(x, ...)

\method{nvertices}{default}(x, ...)

\method{nvertices}{neuronlist}(x, ...)
}
\arguments{
\item{x}{An object with 3d vertices (e.g. neuron, surface etc)}

\item{...}{Additional arguments passed to methods (currently ignored)}
}
\value{
an integer number of vertices (or a vector of length equal to a
  neuronlist)
}
\description{
Find the number of vertices in an object (or each element of a neuronlist)
}
\examples{
nvertices(Cell07PNs[[1]])
nvertices(kcs20)
}
