% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphometry.R
\name{overlap_score}
\alias{overlap_score}
\title{Generate a connectivity matrix based on euclidean distance between points}
\usage{
overlap_score(outputneurons, inputneurons, delta = 1, progress = TRUE)
}
\arguments{
\item{outputneurons}{first set of neurons}

\item{inputneurons}{second set of neurons}

\item{delta}{the distance (in um) at which a synapse might occur}

\item{progress}{whether or not to have a progress bar}
}
\value{
a matrix of overlap scores
}
\description{
Generates an 'overlap matrix' of overlap scores between neurons in the \code{outputneurons} and \code{inputneurons} pools.
For every point in a given neuron in \code{outputneurons}, a distance score is calculated to every point in a neuron in \code{inputneurons}.
The sum of this score is added to the final output matrix. The score is calculated as \code{e(-d^2/(2*delta^2))}, where d is the euclidean distance between the two points,
and delta is the expected distance in um that is considered 'close'. It is recommended that the user resamples neurons before use, using \code{\link{resample}}.
}
\examples{
\dontrun{ 
# Calculate how much some neurons overlap with one another
## Example requires the package nat.flybrains
Cell07PNs_overlap = overlap_score(outputneurons = Cell07PNs, inputneurons = Cell07PNs)

## Plot the results
heatmap(Cell07PNs_overlap)
} 
}
\seealso{
\code{\link{potential_synapses}}, \code{\link{resample}}
}
