% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot3d.boundingbox}
\alias{plot3d.boundingbox}
\title{Plot a bounding box in 3D}
\usage{
\method{plot3d}{boundingbox}(x, ...)
}
\arguments{
  \item{x}{the \code{\link{boundingbox}} object to plot.}

  \item{...}{additional arguments to pass to
  \code{\link[rgl]{segments3d}}.}
}
\value{
A list of RGL object IDs.
}
\description{
Plot a bounding box in 3D
}

