% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hxsurf.R
\name{plot3d.hxsurf}
\alias{plot3d.hxsurf}
\title{Plot amira surface objects in 3D using rgl}
\usage{
\method{plot3d}{hxsurf}(
  x,
  materials = NULL,
  col = NULL,
  gridlines = FALSE,
  ...,
  plotengine = getOption("nat.plotengine")
)
}
\arguments{
\item{x}{An hxsurf surface object}

\item{materials}{Character vector or \code{\link[base]{regex}} naming
materials to plot (defaults to all materials in x). See
\code{\link{subset.hxsurf}}.}

\item{col}{Character vector specifying colors for the materials, or a
function that will be called with the number of materials to plot. When
\code{NULL} (default) will use material colours defined in Amira (if
available), or \code{rainbow} otherwise.}

\item{gridlines}{Whether to display gridlines when using plotly as the backend plotting
engine (default: \code{FALSE})}

\item{...}{Additional arguments passed to \code{triangles3d}}

\item{plotengine}{the plotting backend engine to use either 'rgl' or
'plotly'.}
}
\description{
Plot amira surface objects in 3D using rgl
}
\examples{
plot3d(kcs20)
plot3d(MBL.surf)

\donttest{
# plot only vertical lobe
nclear3d()
plot3d(MBL.surf, materials="VL", alpha=0.3)

# everything except vertical lobe
nclear3d()
plot3d(MBL.surf, alpha=0.3, 
  materials=grep("VL", MBL.surf$RegionList, value = TRUE, invert = TRUE))
}
}
\seealso{
\code{\link{read.hxsurf}}

Other hxsurf:
\code{\link[=as.hxsurf]{as.hxsurf()}},
\code{\link{as.mesh3d}},
\code{\link[=materials]{materials()}},
\code{\link[=read.hxsurf]{read.hxsurf()}},
\code{\link[=subset.hxsurf]{subset.hxsurf()}},
\code{\link[=write.hxsurf]{write.hxsurf()}}
}
\concept{hxsurf}
