% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron-plot.R
\name{plot3d.neuron}
\alias{plot3d.neuron}
\title{Plot neurons in 3D using rgl library or plotly module}
\usage{
\method{plot3d}{neuron}(x, WithLine = TRUE, NeuronNames = FALSE,
  WithNodes = TRUE, WithAllPoints = FALSE, WithText = FALSE,
  PlotSubTrees = TRUE, add = TRUE, col = NULL, soma = FALSE, ...,
  plotengine = getOption("nat.plotengine"))
}
\arguments{
\item{x}{A neuron to plot}

\item{WithLine}{Whether to plot lines for all segments in neuron}

\item{NeuronNames}{Logical indicating whether to label the neuron in the plot
using the NeuronName field \strong{or} a character vector of names.}

\item{WithNodes}{Whether to plot dots for branch and end points}

\item{WithAllPoints}{Whether to plot dots for all points in the neuron}

\item{WithText}{Whether to label plotted points with their numeric id (see
details)}

\item{PlotSubTrees}{Whether to plot all sub trees when the neuron is not
fully connected.}

\item{add}{Whether to add the neuron to existing rgl plot rather than
clearing the scene (default TRUE)}

\item{col}{Colour specification (see rgl materials)}

\item{soma}{Whether to plot a sphere at neuron's origin representing the
soma. Either a logical value or a numeric indicating the radius (default
\code{FALSE}). When \code{soma=TRUE} the radius is hard coded to 2.}

\item{...}{Additional arguments passed to \code{\link[rgl]{lines3d}} (and
\code{\link[rgl]{spheres3d}} if somata are being plotted).}

\item{plotengine}{the plotting backend engine to use either 'rgl' or 'plotly'.}
}
\value{
list of rgl plotting ids (invisibly) separated into \code{lines},
  \code{points}, \code{texts} according to plot element. See
  \code{rgl::\link[rgl]{plot3d}} for details.
}
\description{
Plot neurons in 3D using rgl library or plotly module
}
\details{
Note that when \code{WithText=TRUE}, the numeric identifiers plotted
  are \emph{raw indices} into the \code{x$d} array of the \code{neuron},
  \emph{not} the values of the \code{PointNo} column.

  Note that \code{...} is passed to both \code{\link[rgl]{lines3d}} and
  \code{\link[rgl]{spheres3d}} (if somata are being plotted). Not all
  \code{...} elements are necessarily relevant to both of these drawing
  calls. Furthermore plotting a large number of somata with transparency
  (i.e. \code{alpha < 1} ) can quickly result in very slow rgl draw and
  refresh speeds; you will likely want to set \code{skipRedraw=FALSE} when
  using \code{\link{plot3d.neuronlist}} to plot a collection of neurons.
}
\examples{
# A new plot would have been opened if required
open3d()
plot3d(Cell07PNs[[1]],col='red')
plot3d(Cell07PNs[[2]],col='green')
\donttest{
# clear the current plot
nclear3d()
plot3d(Cell07PNs[[2]],col='blue',add=FALSE)
# plot the number of all nodes
nclear3d()
plot3d(Cell07PNs[[2]],col='red',WithText=TRUE,add=FALSE)
# include cell bodies
plot3d(Cell07PNs[3:4], col='red', soma=TRUE)
plot3d(Cell07PNs[5], col='red', soma=3)
}
}
\seealso{
\code{\link{plot3d.neuronlist}}, \code{\link{plot3d.dotprops}},
  \code{nat::\link[nat]{plot3d}}, \code{rgl::\link[rgl]{plot3d}}
}
