\name{plot3d.neuronlistfh}
\alias{plot3d.neuronlistfh}
\title{plot neurons stored in a neuronlistfh}
\usage{
\method{plot3d}{neuronlistfh}(x, ...)
}
\arguments{
  \item{x}{a neuron list (where omitted will use MyNeurons
  as default)}

  \item{...}{options passed on to plot3d (such as colours,
  line width etc)}
}
\description{
plot neurons stored in a neuronlistfh
}

