% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron-plot.R
\name{plot3d.ngraph}
\alias{plot3d.ngraph}
\title{Plot 3d representation of neuron (ngraph) with directed edges}
\usage{
\method{plot3d}{ngraph}(x, type = "lines", soma = 1, labels = c("none", "nodes", "all"), ...)
}
\arguments{
\item{x}{A \code{\link{ngraph}} object}

\item{type}{They type of arrows (lines by default, see
\code{\link[rgl]{arrow3d}} for details).}

\item{soma}{radius of soma (or \code{FALSE} to suppress plotting)}

\item{labels}{Whether to label nodes/all points with their raw index (not id)}

\item{...}{Additional arguments passed to \code{\link[rgl]{arrow3d}}}
}
\description{
Plot 3d representation of neuron (ngraph) with directed edges
}
\examples{
plot3d(as.ngraph(Cell07PNs[[1]]), labels='nodes')
}
