% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_synapses.R
\name{potential_synapses}
\alias{potential_synapses}
\alias{potential_synapses.neuronlist}
\alias{potential_synapses.neuron}
\alias{potential_synapses.dotprops}
\title{Calculate number of potential synapses between two neurons}
\usage{
potential_synapses(a, b, s, ...)

\method{potential_synapses}{neuronlist}(a, b, s, ...)

\method{potential_synapses}{neuron}(
  a,
  b,
  s,
  sigma = s,
  bounds,
  method = c("direct", "approx"),
  ...
)

\method{potential_synapses}{dotprops}(
  a,
  b,
  s,
  sigma = s,
  seglength = 1,
  bounds = NULL,
  method = c("direct", "approx"),
  ...
)
}
\arguments{
\item{a, b}{neurons or neuronlists}

\item{s}{the approach distance to consider a potential synapse}

\item{...}{Additional arguments passed to methods (see details )}

\item{sigma}{the smoothing parameter in the approximate method (see details)}

\item{bounds}{Optional bounding box to restrict comparison}

\item{method}{Whether to use the direct or approximate method (see details)}

\item{seglength}{how long to consider each distance between points.}
}
\description{
This implements the method of Stepanyants and Chklovskii
}
\details{
Note that \code{potential_synapses.neuronlist} uses 
  \code{\link{nlapply}} to process its first argument (\code{a}). This 
  enables progress bars, robustness to errors and simple parallel execution. 
  See the \code{\link{nlapply}} examples for further details of these 
  arguments in action.
  
  For this reason if you have two neuronlists of unequal sizes, it is 
  recommended to put the larger one in argument \code{a}.
}
\examples{
potential_synapses(Cell07PNs[1], Cell07PNs[1:3], s=2)
\dontrun{
# if you have many neurons to calculate you should get a progress bar 
potential_synapses(Cell07PNs[1:10], Cell07PNs[11:20], s=2)

# you can also use parallel execution, here over 7 cores
# doMC::registerDoMC(7)
potential_synapses(Cell07PNs[1:10], Cell07PNs[11:20], s=2, .parallel=TRUE)
}
}
\references{
Neurogeometry and potential synaptic connectivity. Stepanyants A,
  Chklovskii DB. Trends Neurosci. 2005 Jul;28(7):387-94. 
  \doi{10.1016/j.tins.2005.05.006}
}
\seealso{
Other neuron:
\code{\link[=neuron]{neuron()}},
\code{\link[=ngraph]{ngraph()}},
\code{\link[=plot.dotprops]{plot.dotprops()}},
\code{\link[=prune]{prune()}},
\code{\link[=resample]{resample()}},
\code{\link[=rootpoints]{rootpoints()}},
\code{\link[=spine]{spine()}},
\code{\link[=subset.neuron]{subset.neuron()}}
}
\concept{neuron}
