% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron.R, R/neuronlist.R
\name{prune_twigs}
\alias{prune_twigs}
\alias{prune_twigs.neuron}
\alias{prune_twigs.neuronlist}
\title{Remove all twigs less than a certain path length from a neuron}
\usage{
prune_twigs(x, ...)

\method{prune_twigs}{neuron}(x, twig_length, ...)

\method{prune_twigs}{neuronlist}(x, twig_length, OmitFailures = NA, ...)
}
\arguments{
\item{x}{A \code{\link{neuron}} or \code{\link{neuronlist}} object}

\item{...}{Additional arguments passed to \code{\link{nlapply}},
\code{\link{prune_vertices}} and eventually \code{\link{as.ngraph}}.}

\item{twig_length}{Twigs shorter than this will be pruned}

\item{OmitFailures}{Whether to omit neurons for which \code{FUN} gives an
error. The default value (\code{NA}) will result in nlapply stopping with
an error message the moment there is an error. For other values, see
details.}
}
\description{
\code{prune_twigs} will prune twigs less than a certain path length from a neuron
}
\examples{
# Prune twigs up to 5 microns long
pt5=prune_twigs(Cell07PNs[1:3], twig_length = 5)
# compare original (coloured) and pruned (black) neurons
plot(Cell07PNs[1:3], WithNodes=FALSE, lwd=2, xlim=c(240,300), ylim=c(120, 90))
plot(pt5, WithNodes=FALSE, add=TRUE, lwd=2, col='black')
}
\author{
Gregory Jefferis \email{jefferis@gmail.com}
}
