% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.landmarks}
\alias{read.landmarks}
\title{Generic function to read landmarks in any supported format}
\usage{
read.landmarks(f, ...)
}
\arguments{
  \item{f}{Path to a file (can also be a URL)}

  \item{...}{Additional arguments passed on to format
  specific functions}
}
\description{
Generic function to read landmarks in any supported format
}
\details{
Presently the supported formats are \itemize{

\item Amira

\item CMTK

\item Fiji (see
\url{http://fiji.sc/Name_Landmarks_and_Register}) }
}

