% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.neuron.neuroml}
\alias{read.neuron.neuroml}
\title{Read one or more neurons from a NeuroML v1 file}
\usage{
read.neuron.neuroml(f, ...)
}
\arguments{
  \item{f}{Path to a NeuroML format XML file}

  \item{...}{Additional arguments (presently ignored)}
}
\value{
When the XML file contains only 1 cell, a \code{neuron}
object, otherwise a \code{neuronlist} containing multiple
neurons.
}
\description{
Read one or more neurons from a NeuroML v1 file
}
\references{
\url{http://www.neuroml.org/specifications}
}

