% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglist.R
\name{reglist}
\alias{reglist}
\title{A simple wrapper class for multiple transformations}
\usage{
reglist(..., swap = NULL)
}
\arguments{
\item{...}{One or more transformations}

\item{swap}{A vector of the same length as \code{...} indicating whether the 
direction of each transformation should be swapped (i.e. mapping reference 
-> sample).}
}
\description{
A \code{reglist} is read as a set of transformations to be 
  applied sequentially starting with the first element, then applying the 
  second transformation to the result of the first and so on. Each individual
  transformation is considered to map data from the sample (floating/moving) 
  space to the reference (fixed/template) space.
  
  Each transformation may have an attribute \code{"swap"} indicating that the
  natural direction of the transformation should be swapped (i.e. inverted). 
  This can be done trivially in the case of affine transformations, 
  expensively for others such as CMTK registrations (see 
  \code{\link{cmtkreg}}) and not at all for others. Note that the term 'swap'
  is used to avoid a direct equivalence with inversion - many registration 
  tools use the term \emph{inverse} for directions that one might naively 
  think of as as the natural direction of the transformation (see 
  \code{\link{xformpoints.cmtkreg}} for discussion).
}
\details{
The swap argument is provided as a convenience, but an attribute
  \code{'swap'} can also be set directly on each reigstration.
}
\seealso{
\code{\link{xform}}
}

