% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron.R
\name{reroot}
\alias{reroot}
\alias{reroot.neuron}
\title{Reroot neuron}
\usage{
reroot(x, idx = NULL, point = NULL, pointno = NULL)

\method{reroot}{neuron}(x, idx = NULL, point = NULL, pointno = NULL)
}
\arguments{
\item{x}{A \code{\link{neuron}} or \code{\link{neuronlist}} object}

\item{idx}{index of a node of new soma}

\item{point}{numeric with X,Y,Z coordinates (data.frame or Nx3 matrix for
neuronlist)}

\item{pointno}{new root node id}
}
\value{
neuron with a new soma position
}
\description{
\code{reroot} change soma of a neuron

Change soma node to specific index or a node closes to a specific point.
}
\details{
If both \code{idx} and \code{point} are NULL it returns identity.
}
\examples{
newCell07PN <- reroot(Cell07PNs[[2]], 5)
newCell07PN$StartPoint # 5
}
