% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scale.dotprops}
\alias{scale.dotprops}
\title{Scale and Centre dotprops coords}
\usage{
\method{scale}{dotprops}(x, center = TRUE, scale = TRUE)
}
\arguments{
  \item{x}{A dotprops object}

  \item{center}{3-vector to subtract from x,y,z coords.
  Note that it is possible to scale individual axes}

  \item{scale}{3-vector used to divide x,y,z coords}
}
\value{
neuron with scaled coordinates
}
\description{
Scale and Centre dotprops coords
}
\details{
Note that if scale=TRUE, the neuron will be rescaled to
unit sd in each axis likewise if center=TRUE, the neuron
will be centred around the axis means
}
\seealso{
\code{\link{scale.default}}
}

