% Generated by roxygen2 (4.0.2): do not edit by hand
\name{screen_neurons}
\alias{screen_neurons}
\title{Scan through a set of neurons, plotting each in 3D}
\usage{
screen_neurons(neurons, flip = F, col = "red", Verbose = T, Wait = T,
  sleep = 0.1, extrafun = NULL, selected_file = NULL,
  selected_col = "green", yaml = TRUE, ...)
}
\arguments{
  \item{neurons}{character vector of names of neuron to
  plot.}

  \item{flip}{logical vector indicating whether to use list
  containing flipped neurons. Recycled if length 1.}

  \item{col}{the color with which to plot the neurons
  (default \code{'red'}).}

  \item{Verbose}{logical indicating that info about each
  selected neuron should be printed (default \code{TRUE}).}

  \item{Wait}{logical indicating that there should be a
  pause between each displayed neuron.}

  \item{sleep}{time to pause between each displayed neuron
  when \code{Wait=TRUE}.}

  \item{extrafun}{an optional function called when each
  neuron is plotted, with args \code{gene_name} and
  \code{selected}.}

  \item{selected_file}{an optional path to a \code{yaml}
  file that already contains a selection.}

  \item{selected_col}{the color in which selected neurons
  (such as those specified in \code{selected_file}) should
  be plotted.}

  \item{yaml}{a logical indicating that selections should
  be saved to disk in \code{yaml} rather than \code{rda}
  format.}

  \item{...}{extra arguments to pass to
  \code{\link{plot3d}}.}
}
\value{
A character vector of names of any selected neurons,
\code{NULL} if none selected.
}
\description{
Can also choose to select specific neurons along the way
and navigate forwards and backwards.
}

