% Generated by roxygen2 (4.0.1): do not edit by hand
\name{seglengths}
\alias{seglengths}
\title{Calculate length of all segments in neuron}
\usage{
seglengths(x, all = FALSE, flatten = TRUE)
}
\arguments{
  \item{x}{A neuron}

  \item{all}{Whether to calculate lengths for all segments
  when there are multiple subtrees (default: \code{FALSE})}

  \item{flatten}{Whether to flatten the lists of lists into
  a single list when \code{all=TRUE}}
}
\value{
A vector of lengths for each segment.
}
\description{
Calculate length of all segments in neuron
}
\details{
Only segments in x$SegList will be calculated. Segments
containing only one point will have 0 length.
}
\examples{
summary(seglengths(Cell07PNs[[1]]))
}
\seealso{
\code{\link{as.seglist.neuron}}
}

