% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphometry.R
\name{sholl_analysis}
\alias{sholl_analysis}
\alias{sholl_analysis.neuron}
\alias{sholl_analysis.neuronlist}
\title{Perform a sholl analysis on neuron skeletons}
\usage{
sholl_analysis(x, start = colMeans(xyzmatrix(x)),
  starting.radius = radius.step, ending.radius = 1000,
  radius.step = ending.radius/100)

\method{sholl_analysis}{neuron}(x, start = colMeans(xyzmatrix(x)),
  starting.radius = radius.step, ending.radius = 1000,
  radius.step = ending.radius/100)

\method{sholl_analysis}{neuronlist}(x, start = colMeans(xyzmatrix(x)),
  starting.radius = radius.step, ending.radius = 1000,
  radius.step = ending.radius/100)
}
\arguments{
\item{x}{a neuron or neuronlist object}

\item{start}{the origin from which spheres are grown for the Sholl analysis}

\item{starting.radius}{the radius of the first sphere. Defaults to the radius step}

\item{ending.radius}{the radius of the last sphere. If NULL the distance to the furthest dendritic point from the start point is taken}

\item{radius.step}{the change in radius between successive spheres. Defaults to one 100th of the radius of the ending sphere}
}
\value{
a data.frame of spheres radii and the number of dendritic intersections at each radius
}
\description{
Functions for Sholl analysis of neuronal skeletons
}
\examples{
\dontrun{ 
Cell07PNs_sholl = sholl_analysis(x = Cell07PNs, radius.step = 1, ending.radius = 100)
head(Cell07PNs_sholl[[1]])
} 
}
\seealso{
\code{\link{potential_synapses}}, \code{\link{resample}}, \code{\link{overlap}}
}
