% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron.R
\name{simplify_neuron2}
\alias{simplify_neuron2}
\title{Simplify a neuron to the longest tree with n branch points}
\usage{
simplify_neuron2(x, n = 1, invert = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[nat]{neuron}} to simplify}

\item{n}{Required number of branch points (default=1, minimum 0)}

\item{invert}{Whether to keep the simplified backbone (when
\code{invert=FALSE}, the default) or its inverse.}

\item{...}{Additional arguments (currently ignored)}
}
\description{
Simplify a neuron to the longest tree with n branch points
}
\details{
If the neuron already contains fewer than or exactly the requested
  number of branches, then the original neuron is returned. The approach is 
  basically to compute the longest path in the neuron and then collect that path,
  further delete that path and recompute the longest path and then collect that new
  path again. Perform this approach until you reach the requested number of branchpoints.
}
\examples{
\dontrun{
library(catmaid)
dl1=read.neuron.catmaid(catmaid_skids('annotation:DL1')[1])
dl1.simp=simplify_neuron2(dl1)
dl1.simp4=simplify_neuron2(dl1, n=4)
nclear3d()
plot3d(dl1, col='green', alpha = 0.5, WithNodes = FALSE)
plot3d(dl1.simp4, col='blue', alpha = 0.5, add = TRUE)
plot3d(dl1.simp, col='red', alpha = 0.5, add = TRUE)

# calculate the inverse as well
dl1.simp4.inv=simplify_neuron2(dl1, n=4, invert=TRUE)
nclear3d()
plot3d(dl1.simp4, col='blue',alpha = 0.5)
plot3d(dl1.simp4.inv, col='red', alpha = 0.5, add = TRUE)
}
}
\seealso{
\code{\link[nat]{simplify_neuron}}
}
\author{
Sridhar Jagannathan \email{j.sridharrajan@gmail.com}
}
