% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglist.R
\name{simplify_reglist}
\alias{simplify_reglist}
\title{Simplify a registration list}
\usage{
simplify_reglist(reg, as.cmtk = FALSE)
}
\arguments{
\item{reg}{A registration list (\code{\link{reglist}}) containing one or more
transformations.}

\item{as.cmtk}{Whether to convert to a vector of CMTK format registrations
(see \code{\link{cmtkreg}}). The default value of \code{as.cmtk=NULL} converts}
}
\description{
Simplify a registration list
}
\details{
This function \itemize{
  
  \item inverts any affine matrices with attribute \code{"swap"}
  
  \item collapses multiple affine matrices into a single affine
  
  \item optionally converts all registrations to CMTK on disk registrations 
  when possible.
  
  }
  
  Note that
}
\seealso{
\code{\link{reglist}}, \code{\link{xform}}, \code{\link{cmtkreg}}
}

