% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron.R
\name{stitch_neurons}
\alias{stitch_neurons}
\title{Stitch multiple fragments into single neuron using nearest endpoints}
\usage{
stitch_neurons(x, prefer_soma = TRUE, sort = TRUE, warndist = 1000)
}
\arguments{
\item{x}{A neuronlist containing fragments of a single neuron}

\item{prefer_soma}{When TRUE (the default) the fragment tagged as the soma
will be used as the master neuron.}

\item{sort}{When TRUE (the default) the fragments will be sorted by the
number of nodes they contain before stitching.}

\item{warndist}{If distance is greater than this value, create a warning.}
}
\value{
A single \code{neuron} object containing all input fragments.
}
\description{
Stitch multiple fragments into single neuron using nearest endpoints
}
\details{
Neurons will be ordered by default such the largest (by node count)
  neuron with a soma tag is the \code{master} neuron - i.e. the one
  containing the root node. Fragments are joined recursively in this sort
  order each time simply picking the closest fragment to the current
  \emph{master}. Closest is here defined by the distance between nearest
  endpoints.
}
\examples{
\dontrun{
pn57334=read.neurons.catmaid("name:PN 57334")
# there were 3 fragments when I wrote this that all contained PN 57334 in
# their name
length(pn57334)
summary(pn57334)
pn57334.whole=stitch_neurons(pn57334)
summary(pn57334.whole)
plot3d(pn57334.whole)
}
}
\seealso{
\code{\link{stitch_neuron}}
}
\author{
Gregory Jefferis \email{jefferis@gmail.com}
}
