% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron.R
\name{stitch_neurons_mst}
\alias{stitch_neurons_mst}
\title{Stitch multiple fragments into single neuron using minimum spanning tree}
\usage{
stitch_neurons_mst(x)
}
\arguments{
\item{x}{Fragments that could be neuronlist or a single neuron with multiple trees(fragments)}
}
\value{
A single \code{neuron} object containing all input fragments.
}
\description{
Stitch multiple fragments into single neuron using minimum spanning tree
}
\details{
The neurons are joined using the minimum spanning tree which the tree which minimises the sum of
edge weights (here they would be euclidean distance). The neuron is also rooted considering the root of the 
largest cluster.
}
\examples{
\dontrun{
library(catmaid)
dl1=read.neuron.catmaid(catmaid_skids('annotation:DL1')[1])
dl1_branches=simplify_neuron(dl1, n = 1, invert = T)
#Example with a single neuron, that has several fragments which could be stiched together..
dl1_branches$nTrees
dl1_whole=stitch_neurons_mst(dl1_branches)
plot3d(dl1_whole)
#Example with neuronlist
dl1_main=simplify_neuron(dl1, n = 1, invert = F)
dl1_branches=simplify_neuron(dl1, n = 1, invert = T)
dl1_fragment <- list(dl1_main,dl1_branches)
dl1_fragment <- as.neuronlist(dl1_fragment)
dl1_whole = stitch_neurons_mst(dl1_fragment)

}
}
\seealso{
\code{\link{simplify_neuron}}
}
\author{
Sridhar Jagannathan \email{j.sridharrajan@gmail.com}
}
