% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{sub2ind}
\alias{sub2ind}
\title{Find 1D index given n-dimensional indices}
\usage{
sub2ind(dims, indices)
}
\arguments{
\item{dims}{vector of dimensions of object to index into.}

\item{indices}{vector of n-dimensional indices.}
}
\description{
Emulates the MATLAB function \code{sub2ind}.
}
\details{
There is a *much* more efficient version implemented in natcpp > 0.1.1
}
\examples{
dims <- 3:5
ijk <- matrix(c(1L, 1L, 1L, 3L, 4L, 5L), ncol = 3, byrow = TRUE)
sub2ind(dims, ijk)
}
