% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hxsurf.R
\name{subset}
\alias{subset}
\title{Subset methods for different nat objects}
\description{
These methods enable subsets of some nat objects including neurons and
neuronlists to be obtained. See the help for each individual method for
details.
}
\seealso{
\code{\link{subset.neuron}}, \code{\link{subset.dotprops}},
  \code{\link{subset.hxsurf}}, \code{\link{subset.neuronlist}}
}

