% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.neuron}
\alias{summary.dotprops}
\alias{summary.neuron}
\title{Summary statistics for individual neurons}
\usage{
\method{summary}{neuron}(object, ...)

\method{summary}{dotprops}(object, veclength = 1, ...)
}
\arguments{
\item{object}{The \code{neuron} to summarise}

\item{...}{Additional arguments (currently ignored)}

\item{veclength}{The vector length to assume for each segment so that a cable
length estimate can be made.}
}
\description{
Summary statistics for individual neurons
}
\examples{
summary(Cell07PNs[[1]])
summary(kcs20[[1]])
}

