% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.neuronlist}
\alias{summary.neuronlist}
\title{Summary statistics for neurons in a neuronlist (e.g. cable length, nodes)}
\usage{
\method{summary}{neuronlist}(object)
}
\arguments{
\item{object}{The \code{neuronlist} to summarise}

\item{...}{Additional arguments to summary methods for neuronlist objects}
}
\description{
Summary statistics for neurons in a neuronlist (e.g. cable length, nodes)
}
\examples{
summary(Cell07PNs)
}

