% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire3d.R
\name{wire3d.mesh3d}
\alias{wire3d.mesh3d}
\title{methods here imported from rgl}
\usage{
\method{wire3d}{mesh3d}(
  x,
  override = TRUE,
  meshColor = c("vertices", "edges", "faces", "legacy"),
  ...
)
}
\arguments{
\item{x}{a \code{mesh3d} object (class \code{qmesh3d} or \code{tmesh3d}).}

\item{override}{should the parameters specified here override those stored in the object?}

\item{meshColor}{how should colours be interpreted?  See details 
  below}

\item{...}{additional rendering parameters}
}
\description{
methods here imported from rgl
}
