% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire3d.R
\name{wire3d.shapelist3d}
\alias{wire3d.shapelist3d}
\title{methods here imported from rgl}
\usage{
\method{wire3d}{shapelist3d}(x, override = TRUE, ...)
}
\arguments{
\item{x}{object of type 'shapelist3d'}

\item{override}{Whether the material properties should override the ones in the shapes}

\item{...}{Additional arguments passed to \code{\link[rgl]{wire3d}}}
}
\description{
methods here imported from rgl
}
