% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmtk_io.R
\name{write.cmtk}
\alias{write.cmtk}
\title{Write a suitable list to a CMTK TypedStream file on disk}
\usage{
write.cmtk(l, con, gzip = FALSE, version = NA_character_)
}
\arguments{
\item{l}{Appropriately formatted list}

\item{con}{A character string specifying a path or a connection}

\item{gzip}{Whether to gzip output file (default FALSE)}

\item{version}{TYPEDSTREAM version number, defaults to \code{"1.1"} if not 
specified in the version attribute of \code{l}.}
}
\description{
This is probably only of interest to developers. End users will 
  probably wish to use more specific functions such as write.cmtkreg for
  writing out registrations.
}
\details{
NB a version specified on the command line overrides one encoded as 
  an attribute in the input list.
}
\seealso{
Other cmtk-io:
\code{\link[=cmtk.extract_affine]{cmtk.extract_affine()}},
\code{\link[=read.cmtk]{read.cmtk()}},
\code{\link[=read.cmtkreg]{read.cmtkreg()}},
\code{\link[=write.cmtkreg]{write.cmtkreg()}}
}
\concept{cmtk-io}
