% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hxsurf.R
\name{write.hxsurf}
\alias{write.hxsurf}
\title{Write Amira surface (aka HxSurface or HyperSurface) into .surf file.}
\usage{
write.hxsurf(surf, filename)
}
\arguments{
\item{surf}{hxsurf object to write to file.}

\item{filename}{character vector defining path to file.}
}
\value{
\code{NULL} or integer status from \code{\link{close}}.
}
\description{
Write Amira surface (aka HxSurface or HyperSurface) into .surf file.
}
\seealso{
\code{\link{plot3d.hxsurf}},\code{\link{read.hxsurf}}, \code{\link{rgb}}

Other amira:
\code{\link[=amiratype]{amiratype()}},
\code{\link[=is.amiramesh]{is.amiramesh()}},
\code{\link[=read.amiramesh]{read.amiramesh()}},
\code{\link[=read.hxsurf]{read.hxsurf()}}

Other hxsurf:
\code{\link[=as.hxsurf]{as.hxsurf()}},
\code{\link{as.mesh3d}},
\code{\link[=materials]{materials()}},
\code{\link[=plot3d.hxsurf]{plot3d.hxsurf()}},
\code{\link[=read.hxsurf]{read.hxsurf()}},
\code{\link[=subset.hxsurf]{subset.hxsurf()}}
}
\concept{amira}
\concept{hxsurf}
