% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmtk_geometry.R
\name{affmat2cmtkparams}
\alias{affmat2cmtkparams}
\title{Decompose homogeneous affine matrix to CMTK registration parameters}
\usage{
affmat2cmtkparams(matrix, centre = c(0, 0, 0))
}
\arguments{
\item{matrix}{4x4 homogeneous affine matrix}

\item{centre}{Rotation centre}
}
\value{
5x3 matrix of CMTK registration parameters with a version attribute
}
\description{
Decompose homogeneous affine matrix to CMTK registration parameters
}
\details{
The version attribute of the resultant matrix marks this as
  compliant with CMTK>v2.4 (~ Dec 2013) when a bug in affine matrix
  (de)composition was fixed.
}
\seealso{
Other cmtk-geometry: \code{\link{cmtk.dof2mat}},
  \code{\link{cmtk.mat2dof}},
  \code{\link{cmtkparams2affmat}}
}
