% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hxsurf.R
\name{as.mesh3d}
\alias{as.mesh3d}
\alias{as.mesh3d.hxsurf}
\title{Convert an object to an rgl mesh3d}
\usage{
as.mesh3d(x, ...)

\method{as.mesh3d}{hxsurf}(x, Regions = NULL, material = NULL,
  drop = TRUE, ...)
}
\arguments{
  \item{x}{Object to convert to mesh3d}

  \item{...}{Additional arguments for methods}

  \item{Regions}{Character vector or regions to select from
  \code{hxsurf} object}

  \item{material}{rgl materials such as \code{color}}

  \item{drop}{Whether to drop unused vertices (default
  TRUE)}
}
\description{
Note that this provides a link to the Rvcg package
}
\seealso{
\code{\link[rgl]{tmesh3d}}

Other hxsurf: \code{\link{hxsurf}},
\code{\link{read.hxsurf}}; \code{\link{materials}},
\code{\link{materials.character}},
\code{\link{materials.default}},
\code{\link{materials.hxsurf}};
\code{\link{plot3d.hxsurf}}; \code{\link{subset.hxsurf}};
\code{\link{write.hxsurf}}
}

