% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coord2ind}
\alias{coord2ind}
\alias{coord2ind.default}
\title{Find 1D indices into a 3D image given spatial coordinates}
\usage{
coord2ind(coords, ...)

\method{coord2ind}{default}(coords, imdims, voxdims = NULL, origin = NULL,
  aperm, Clamp = FALSE, CheckRanges = !Clamp, ...)
}
\arguments{
\item{coords}{spatial coordinates of image voxels.}

\item{...}{extra arguments passed to methods.}

\item{imdims}{array dimensions of 3D image.}

\item{voxdims}{vector of 3 voxels dimensions (width, height, depth).}

\item{origin}{the origin of the 3D image.}

\item{aperm}{permutation order for axes.}

\item{Clamp}{???}

\item{CheckRanges}{whether to check if coordinates are out of range.}
}
\description{
Find 1D indices into a 3D image given spatial coordinates
}
\seealso{
\code{\link{ind2coord}}, \code{\link{sub2ind}}, \code{\link{ijkpos}}
}

