% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotprops.R
\name{*.dotprops}
\alias{*.dotprops}
\alias{+.dotprops}
\alias{-.dotprops}
\alias{/.dotprops}
\title{Arithmetic for dotprops objects}
\usage{
\method{*}{dotprops}(x, y)

\method{+}{dotprops}(x, y)

\method{-}{dotprops}(x, y)

\method{/}{dotprops}(x, y)
}
\arguments{
\item{x}{A dotprops object}

\item{y}{A scalar or 3-vector that will be applied to the dotprops object}
}
\value{
A new dotprops object
}
\description{
Arithmetic for dotprops objects
}
