% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hxsurf.R
\name{c.hxsurf}
\alias{c.hxsurf}
\title{Concatenate two HyperSurface objects}
\usage{
\method{c}{hxsurf}(x, y)
}
\arguments{
\item{x}{hxsurf}

\item{y}{another hxsurf}
}
\value{
new hxsurf
}
\description{
Concatenate two HyperSurface objects
}
\examples{
h1 = as.hxsurf(icosahedron3d(), 'a')
h2 = as.hxsurf(tetrahedron3d(), 'b')
h3=c(h1, h2)
}
