\name{imexpand.grid}
\alias{imexpand.grid}
\title{Convert locations of im3d voxel grid into XYZ coordinates}
\usage{
imexpand.grid(d)
}
\arguments{
  \item{d}{An \code{im3d} object}
}
\value{
Nx3 matrix of image coordindates
}
\description{
Convert locations of im3d voxel grid into XYZ coordinates
}
\seealso{
expand.grid
}

