% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im3d.R
\name{imslice}
\alias{imslice}
\title{Slice out a 3D subarray (or 2d matrix) from a 3D image array}
\usage{
imslice(x, slice, slicedim = "z", drop = TRUE)
}
\arguments{
\item{x}{An im3d objet}

\item{slice}{Indices defining the slices to keep}

\item{slicedim}{Character vector or integer defining axis from which slices 
will be removed.}

\item{drop}{Whether singleton dimensions will be dropped (default: TRUE) 
conveting 3D array to 2d matrix.}
}
\description{
Slice out a 3D subarray (or 2d matrix) from a 3D image array
}
\details{
Note the sample locations stored in the x,y,z attributes will be 
  updated appropriately. FIXME: Should we also update bounding box?
}
\seealso{
Other im3d: 
\code{\link{as.im3d}()},
\code{\link{boundingbox}()},
\code{\link{im3d-coords}},
\code{\link{im3d-io}},
\code{\link{im3d}()},
\code{\link{imexpand.grid}()},
\code{\link{is.im3d}()},
\code{\link{mask}()},
\code{\link{origin}()},
\code{\link{projection}()},
\code{\link{threshold}()},
\code{\link{unmask}()},
\code{\link{voxdims}()}
}
\concept{im3d}
