% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{intersect_plane}
\alias{intersect_plane}
\alias{intersect_plane.default}
\alias{intersect_plane.neuron}
\title{Find the points on a plane that are intersected by an object}
\usage{
intersect_plane(x, plane, ...)

\method{intersect_plane}{default}(x, plane, ...)

\method{intersect_plane}{neuron}(x, plane, closestpoint = NULL, ...)
}
\arguments{
\item{x}{A neuron, set of line segments or other data - see details.}

\item{plane}{A plane, specified by the 4 coefficients of the plane equation 
(see \code{\link{plane_coefficients}})}

\item{...}{Additional arguments passed to methods}

\item{closestpoint}{Used to define the closest hit when there are multiple}
}
\value{
A Nx3 matrix of the X,Y,Z positions of the intersections (NA when
  there is no intersection)
}
\description{
\code{intersect_plane.neuron} finds the place where a neuron intersection
}
\examples{
## Find plane coefficients
# point on plane
cent=c(250.4987, 95.73561, 140.2052)
# vector normal to plane
vec=c(0.7709581, 0.03417276, -0.411977)
plc=plane_coefficients(cent, vec)

## intersect with plane
ip=intersect_plane(Cell07PNs[[1]], plc)
plot(Cell07PNs[[1]], WithNodes=FALSE)
points(ip[1], ip[2], pch=19, cex=2, col='red')

\dontrun{
plot3d(Cell07PNs[[1]], col='grey', WithNodes=FALSE)
spheres3d(matrix(ip, ncol=3), col='red', rad=2)
planes3d(plc[,1:3], d=plc[,'d'])
}
}
\seealso{
Other geometry: \code{\link{plane_coefficients}}
}
\concept{geometry}
