% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amiramesh-io.R
\name{is.amiramesh}
\alias{is.amiramesh}
\title{Check if file is amiramesh format}
\usage{
is.amiramesh(f = NULL, bytes = NULL)
}
\arguments{
\item{f}{Path to one or more files to be tested \strong{or} an array of raw 
bytes, for one file only.}

\item{bytes}{optional raw vector of at least 11 bytes from the start of a 
single file (used in preference to reading file \code{f}).}
}
\value{
logical
}
\description{
Check if file is amiramesh format
}
\details{
Tries to be as fast as possible by reading only first 11 bytes and 
  checking if they equal to "# AmiraMesh" or (deprecated) "# HyperMesh".
}
\seealso{
Other amira: \code{\link{amiratype}},
  \code{\link{read.amiramesh}}, \code{\link{read.hxsurf}},
  \code{\link{write.hxsurf}}
}
