% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im3d.R
\name{is.im3d}
\alias{is.im3d}
\title{Test if an object is of class im3d}
\usage{
is.im3d(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
logical
}
\description{
Test if an object is of class im3d
}
\seealso{
Other im3d: \code{\link{as.im3d}},
  \code{\link{boundingbox}}, \code{\link{im3d-coords}},
  \code{\link{im3d-io}}, \code{\link{im3d}},
  \code{\link{imexpand.grid}}, \code{\link{imslice}},
  \code{\link{mask}}, \code{\link{origin}},
  \code{\link{projection}}, \code{\link{threshold}},
  \code{\link{unmask}}, \code{\link{voxdims}}
}
