% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/im3d.R
\name{is.im3d}
\alias{is.im3d}
\title{Test if an object is of class im3d}
\usage{
is.im3d(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
logical
}
\description{
Test if an object is of class im3d
}
\seealso{
Other im3d: \code{\link{as.im3d}},
  \code{\link{as.im3d.im3d}}, \code{\link{as.im3d.matrix}};
  \code{\link{boundingbox}},
  \code{\link{boundingbox.character}},
  \code{\link{boundingbox.default}},
  \code{\link{boundingbox.im3d}},
  \code{\link{boundingbox.list}},
  \code{\link{boundingbox.shape3d}},
  \code{\link{boundingbox<-}}; \code{\link{ijkpos}},
  \code{\link{im3d-coords}}, \code{\link{xyzpos}};
  \code{\link{im3d-io}}, \code{\link{read.im3d}},
  \code{\link{write.im3d}}; \code{\link{im3d}};
  \code{\link{imexpand.grid}}; \code{\link{imslice}};
  \code{\link{origin}}; \code{\link{projection}};
  \code{\link{threshold}}, \code{\link{threshold.im3d}};
  \code{\link{unmask}}; \code{\link{voxdims}},
  \code{\link{voxdims.default}}
}

