% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronlist.R
\name{is.neuronlist}
\alias{is.neuronlist}
\title{Test objects of neuronlist class to store multiple neurons}
\usage{
is.neuronlist(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
A logical indicating whether the object is a neuronlist.
}
\description{
Tests if object is a neuronlist.
}
\details{
\code{is.neuronlist} uses a relaxed definition to cope with older 
  lists of neurons that do not have a class attribute of neuronlist.
}
\seealso{
Other neuronlist: \code{\link{*.neuronlist}},
  \code{\link{neuronlist-dataframe-methods}},
  \code{\link{neuronlistfh}}, \code{\link{neuronlist}},
  \code{\link{nlapply}}, \code{\link{read.neurons}},
  \code{\link{write.neurons}}
}

