% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{manually_assign_axon_dendrite}
\alias{manually_assign_axon_dendrite}
\alias{manually_assign_axon_dendrite.neuron}
\alias{manually_assign_axon_dendrite.neuronlist}
\title{Manually assign a dendrite and axon to a neuron}
\usage{
manually_assign_axon_dendrite(x, soma = FALSE)

\method{manually_assign_axon_dendrite}{neuron}(x, soma = FALSE)

\method{manually_assign_axon_dendrite}{neuronlist}(x, soma = FALSE)
}
\arguments{
\item{x}{a neuron/neuronlist object}

\item{soma}{whether or not to plot a soma, and at what radius}
}
\value{
The neuron/neuronlist object with axon/dendrite info assigned in SWC format to neuron$d
#' @seealso \code{\link{as.neuron.ngraph}}, \code{\link{subset.neuron}}, 
  \code{\link{prune.neuron}}, \code{\link{prune_online}}
}
\description{
Manually assign the dendrite and axon to neurons / a neuron
}
\examples{
\dontrun{ 
## Interactively shoose which bit of the neuron you wish to keep
split.as.you.like.it = manually_assign_axon_dendrite(Cell07PNs)
} 
}
