% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat-data.R
\docType{data}
\name{MBL.surf}
\alias{MBL.surf}
\title{Surface object (hxsurf) for the left mushroom body in FCWB template space}
\description{
This surface object is in the same space as the 20 Kenyon cells in
\code{\link{kcs20}}.
}
\examples{
\donttest{
plot3d(kcs20)
plot3d(MBL.surf, alpha=0.3)
}

\dontrun{
## originally generated as follows
library(nat.flybrains)
MBL.surf=subset(FCWBNP.surf, "MB.*_L", drop = T)
}
}
\seealso{
\code{\link{hxsurf}}

Other nat-data: \code{\link{Cell07PNs}},
  \code{\link{kcs20}}
}
