\docType{package}
\name{nat-package}
\alias{nat}
\alias{nat-package}
\title{Analyse 3D biological image data especially neurons}
\description{
Key functions that you will want to look at include
read.neuron write.neuron
}
\details{
Objects of class neuron neuronlist

plot.neuron plot3d.neuron
}
\seealso{
\code{\link{rgl}} which is used for visualisation.
}
\keyword{package}

