% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron-plot.R
\name{nclear3d}
\alias{nclear3d}
\title{Clear the rgl or plotly 3D scene}
\usage{
nclear3d(plotengine = getOption("nat.plotengine"), ...)
}
\arguments{
\item{plotengine}{the plotting backend engine to use either 'rgl' or 'plotly'.}

\item{...}{Additional arguments passed to
\code{\link[rgl:clear3d]{rgl::clear3d}}}
}
\description{
Clear the rgl or plotly 3D scene
}
\details{
\code{\link{rgl}} and \code{\link[plotly:plot_ly]{plotly}} have
  quite different models for how to handle the active plot. \code{nclear3d}
  allows you to treat them more similarly
}
\examples{

\donttest{
nclear3d()
plot3d(Cell07PNs[[1]])
}
}
\seealso{
\code{\link[rgl:clear3d]{rgl::clear3d}}, \code{\link[nat]{plot3d}},
  \code{\link[nat]{plot3d.neuronlist}}
}
