% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuronlist.R
\name{*.neuronlist}
\alias{*.neuronlist}
\alias{+.neuronlist}
\alias{-.neuronlist}
\alias{/.neuronlist}
\title{Arithmetic for neuron coordinates applied to neuronlists}
\usage{
\method{*}{neuronlist}(x, y)

\method{+}{neuronlist}(x, y)

\method{-}{neuronlist}(x, y)

\method{/}{neuronlist}(x, y)
}
\arguments{
\item{x}{a neuronlist}

\item{y}{(a numeric vector to multiply coords in neuronlist members)}
}
\value{
modified neuronlist
}
\description{
If x is one number or 3-vector, multiply coordinates by that
If x is a 4-vector, multiply xyz and diameter
TODO Figure out how to document arithemtic functions in one go
}
\examples{
mn2<-Cell07PNs[1:10]*2
}
\seealso{
Other neuronlist: \code{\link{[.neuronlist}},
  \code{\link{[<-.neuronlist}}, \code{\link{droplevels}},
  \code{\link{droplevels.neuronlist}}, \code{\link{head}},
  \code{\link{head.neuronlist}},
  \code{\link{neuronlist-dataframe-methods}},
  \code{\link{neuronlist-dataframe-methods}},
  \code{\link{tail}}, \code{\link{tail.neuronlist}},
  \code{\link{with}}, \code{\link{with.neuronlist}};
  \code{\link{as.neuronlistfh}},
  \code{\link{as.neuronlistfh.neuronlist}},
  \code{\link{is.neuronlistfh}},
  \code{\link{neuronlistfh}}, \code{\link{neuronlistfh}};
  \code{\link{is.neuronlist}}; \code{\link{neuronlist}};
  \code{\link{nlapply}}, \code{\link{nmapply}};
  \code{\link{read.neurons}}; \code{\link{write.neurons}}
}

