% Generated by roxygen2 (4.0.1): do not edit by hand
\name{*.neuronlist}
\alias{*.neuronlist}
\alias{+.neuronlist}
\alias{-.neuronlist}
\alias{/.neuronlist}
\title{Arithmetic for neuron coordinates applied to neuronlists}
\usage{
\method{*}{neuronlist}(x, y)

\method{+}{neuronlist}(x, y)

\method{-}{neuronlist}(x, y)

\method{/}{neuronlist}(x, y)
}
\arguments{
  \item{x}{a neuronlist}

  \item{y}{(a numeric vector to multiply coords in
  neuronlist members)}
}
\value{
modified neuronlist
}
\description{
If x is one number or 3-vector, multiply coordinates by
that If x is a 4-vector, multiply xyz and diameter TODO
Figure out how to document arithemtic functions in one go
}
\examples{
mn2<-Cell07PNs[1:10]*2
}

