% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuronlist.R
\name{neuronlist-dataframe-methods}
\alias{droplevels}
\alias{droplevels.neuronlist}
\alias{head}
\alias{head.neuronlist}
\alias{neuronlist-dataframe-methods}
\alias{tail}
\alias{tail.neuronlist}
\alias{with}
\alias{with.neuronlist}
\title{Methods for working with the dataframe attached to a neuronlist}
\usage{
\method{droplevels}{neuronlist}(x, except, ...)

\method{with}{neuronlist}(data, expr, ...)

\method{head}{neuronlist}(x, ...)

\method{tail}{neuronlist}(x, ...)
}
\arguments{
\item{x}{A neuronlist object}

\item{except}{indices of columns from which \emph{not} to drop levels}

\item{...}{Further arguments passed to default methods (and usually ignored)}

\item{data}{A neuronlist object}

\item{expr}{The expression to evaluate}
}
\value{
the attached dataframe with levels dropped (NB \strong{not} the
  neuronlist)
}
\description{
\code{droplevels} Remove redundant factor levels in dataframe
  attached to neuronlist

\code{with} Evaluate expression in the context of dataframe
  attached to a neuronlist

\code{head} Return the first part of data.frame attached to
  neuronlist

\code{tail} Return the last part of data.frame attached to
  neuronlist
}
\seealso{
\code{\link{droplevels}}

\code{\link{with}}

\code{\link{head}}

\code{\link{tail}}
}

