\name{neuronlist}
\alias{neuronlist}
\title{neuronlist class to store multiple neurons}
\description{
  neuronlist objects consist of a list of neuron objects
  along with an optional attached dataframe containing
  information about the neurons. Relevant functions include
  plot3d.neuronlist write.neuronlist subset.neuronlist
  read.neurons can generate a neuronlist
}

