\name{neuronlistfh}
\alias{neuronlistfh}
\title{neuronlistfh class to store multiple neurons cached on disk}
\description{
neuronlistfh objects consist of a list of neuron objects
along with an optional attached dataframe containing
information about the neurons. Relevant functions include
plot3d.neuronlistfh
}
\seealso{
neuronlist
}

