% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronlistz.R
\name{neuronlistz}
\alias{neuronlistz}
\alias{as.neuronlist.neuronlistz}
\title{A neuronlist object that will read neurons from a zip file on demand}
\usage{
neuronlistz(zip, patt = NULL, df = NULL, ...)

\method{as.neuronlist}{neuronlistz}(l, ...)
}
\arguments{
\item{zip}{Path to the zip file}

\item{patt}{Optional regex pattern or function to specify a subset of files.}

\item{df}{A \code{data.frame} of metadata that will be attached to the
\code{neuronlistz} and will define the order of the objects inside it.}

\item{...}{Additional arguments currently ignored}

\item{l}{An existing list or a single neuron to start a list}
}
\value{
A \code{\link{neuronlist}} object with additional class
  \code{neuronlistz}
}
\description{
\code{as.neuronlist.neuronlistz} converts a \code{neuronlistz}
  to a regular (in memory) \code{\link{neuronlist}}
}
\details{
\code{\link{neuronlistz}} is designed to wrap zip files containing
  neurons saved in the RDS or faster/smaller qs format for rapid access. You
  should be able to read typical files in <20 ms. For files of ~3 GB there is
  a fixed cost of the order of 10-15ms per read.
}
\examples{
\donttest{
write.neurons(Cell07PNs[1:5], tf <- tempfile(fileext = '.zip'), format='rds')
nz=neuronlistz(tf)
nz[[1]]
nz[1:5]
}
\dontrun{
write.neurons(Cell07PNs[1:5], tf <- tempfile(fileext = '.zip'), format='qs')
nz2=neuronlistz(tf)
all.equal(nz2[1:3], nz[1:3])
}
}
\seealso{
\code{\link{neuronlist}}, \code{\link{neuronlistfh}},
  \code{\link{write.neurons}}

Other neuronlist: 
\code{\link{*.neuronlist}()},
\code{\link{is.neuronlist}()},
\code{\link{neuronlist-dataframe-methods}},
\code{\link{neuronlistfh}()},
\code{\link{neuronlist}()},
\code{\link{nlapply}()},
\code{\link{read.neurons}()},
\code{\link{write.neurons}()}
}
\concept{neuronlist}
