% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronlist.R
\name{nlapply}
\alias{nlapply}
\alias{progress_natprogress}
\alias{nmapply}
\title{lapply and mapply for neuronlists (with optional parallelisation)}
\usage{
nlapply(
  X,
  FUN,
  ...,
  subset = NULL,
  OmitFailures = NA,
  withMeta = FALSE,
  .progress = getOption("nat.progress", default = "auto")
)

progress_natprogress(...)

nmapply(
  FUN,
  X,
  ...,
  MoreArgs = NULL,
  SIMPLIFY = FALSE,
  USE.NAMES = TRUE,
  subset = NULL,
  OmitFailures = NA,
  .progress = getOption("nat.progress", default = "auto")
)
}
\arguments{
\item{X}{A neuronlist}

\item{FUN}{Function to be applied to each element of X}

\item{...}{Additional arguments for FUN (see details)}

\item{subset}{Character, numeric or logical vector specifying on which subset
of \code{X} the function \code{FUN} should be applied. Elements outside the
subset are passed through unmodified.}

\item{OmitFailures}{Whether to omit neurons for which \code{FUN} gives an
error. The default value (\code{NA}) will result in nlapply stopping with
an error message the moment there is an error. For other values, see
details.}

\item{withMeta}{whether or not to include the metadata from the neuronlist
data.frame}

\item{.progress}{Character vector specifying the type of progress bar (see
Progress bar section for details) The default value of \code{"auto"} shows
a progress bar in interactive use after 2s. The default value can be
overridden for the current session by setting the value of
\code{options(nat.progressbar)} (see examples). Values of \code{T} and
\code{F} are aliases for 'text' and 'none', respectively.}

\item{MoreArgs}{a list of other arguments to \code{FUN}.}

\item{SIMPLIFY}{logical or character string; attempt to reduce the
    result to a vector, matrix or higher dimensional array; see
    the \code{simplify} argument of \code{\link[base]{sapply}}.}

\item{USE.NAMES}{logical; use the names of the first \dots argument, or
    if that is an unnamed character vector, use that vector as the names.}
}
\value{
A neuronlist
}
\description{
Versions of lapply and mapply that look after the class and
  attached dataframe of neuronlist objects. \code{nlapply} can apply a
  function to only a \code{subset} of elements in the input neuronlist.
  Internally \code{nlapply} uses \code{plyr::llply} thereby enabling progress
  bars and simple parallelisation (see plyr section and examples).

\code{progress_natprogress} provides a progress bar compatible
  with the \code{progress::\link[progress]{progress_bar}}.
}
\details{
When \code{OmitFailures} is not \code{NA}, \code{FUN} will be
  wrapped in a call to \code{try} to ensure that failure for any single
  neuron does not abort the nlapply/nmapply call. When
  \code{OmitFailures=TRUE} the resultant neuronlist will be subsetted down to
  return values for which \code{FUN} evaluated successfully. When
  \code{OmitFailures=FALSE}, "try-error" objects will be left in place. In
  either of the last 2 cases error messages will not be printed because the
  call is wrapped as \code{try(expr, silent=TRUE)}.
}
\section{plyr}{
 The arguments of most interest from plyr are:

  \itemize{

  \item \code{.inform} set to \code{TRUE} to give more informative error
  messages that should indicate which neurons are failing for a given applied
  function.

  \item \code{.progress} set to \code{"text"} for a basic progress bar

  \item \code{.parallel} set to \code{TRUE} for parallelisation after
  registering a parallel backend (see below).

  \item \code{.paropts} Additional arguments for parallel computation. See
  \code{\link[plyr]{llply}} for details.

  }

  Before using parallel code within an R session you must register a suitable
  parallel backend. The simplest example is the multicore option provided by
  the \code{doMC} package that is suitable for a spreading computational load
  across multiple cores on a single machine. An example is provided below.

  Note that the progress bar and parallel options cannot be used at the same
  time. You may want to start a potentially long-running job with the
  progress bar option and then abort and re-run with \code{.parallel=TRUE} if
  it looks likely to take a very long time.
}

\section{Progress bar}{
 There are currently two supported approaches to
  defining progress bars for \code{nlapply}. The default (when
  \code{progress="auto"}) now uses a progress bar built using
  \code{\link[progress]{progress_bar}} from the progress package, which can
  be highly customised. The alternative is to use the progress bars
  distributed directly with the \code{plyr} package such as
  \code{\link[plyr]{progress_text}}.

  In either case the value of the \code{.progress} argument must be a
  character vector which names a function. According to \code{plyr}'s
  convention an external function called \code{progress_myprogressbar} will
  be identified by setting the argument to \code{.progress="myprogressbar"}.
  By default the supplied \code{progress_natprogress} function will be used
  when \code{.progress="auto"}; this function will probably not be used
  directly by end users; however it must be exported for \code{nlapply} with
  progress to work properly in other functions.

  For \code{nmapply} only the default \code{nat_progress} bar can be shown
  for architectural reasons. It will be shown in interactive mode when
  \code{.progress='auto'} (the default). The progress bar can be suppressed
  by setting \code{.progress='none'}. Any other value will result in a
  progress bar being shown in both interactive and batch modes.
}

\examples{
## nlapply example
kcs.reduced=nlapply(kcs20,function(x) subset(x,sample(nrow(x$points),50)))
open3d()
plot3d(kcs.reduced,col='red', lwd=2)
plot3d(kcs20,col='grey')
close3d()

\dontrun{
# example of using plyr's .inform argument for debugging error conditions
xx=nlapply(Cell07PNs, prune_strahler)
# oh dear there was an error, let's get some details about the neuron
# that caused the problem
xx=nlapply(Cell07PNs, prune_strahler, .inform=TRUE)
}

\dontrun{
## nlapply example with plyr
## dotprops.neuronlist uses nlapply under the hood
## the .progress and .parallel arguments are passed straight to
system.time(d1<-dotprops(kcs20,resample=1,k=5,.progress='text'))
## plyr+parallel
library(doMC)
# can also specify cores e.g. registerDoMC(cores=4)
registerDoMC()
system.time(d2<-dotprops(kcs20,resample=1,k=5,.parallel=TRUE))
stopifnot(all.equal(d1,d2))
}

## nmapply example
# flip first neuron in X, second in Y and 3rd in Z
xyzflip=nmapply(mirror, kcs20[1:3], mirrorAxis = c("X","Y","Z"),
  mirrorAxisSize=c(400,20,30))

# this artificial example will show a progress bar in interactive use
xyzflip=nmapply(function(...) {Sys.sleep(.2);mirror(...)}, kcs20,
  mirrorAxis= sample(LETTERS[24:26], size = 20, replace = TRUE),
  mirrorAxisSize=runif(20, min=40, max=200))
\donttest{
open3d()
plot3d(kcs20[1:3])
plot3d(xyzflip)
close3d()
}

\dontrun{
## Override default progress bar behaviour via options
# sleep 50ms per neuron to ensure progress bar gets triggered
sl=nlapply(Cell07PNs, FUN = function(x) {Sys.sleep(0.05);seglengths(x)})
# the default progess bar for nat < 1.9.1
options(nat.progress='traditional')
sl=nlapply(Cell07PNs, FUN = seglengths)
# no progress bar ever
options(nat.progress='none')
sl=nlapply(Cell07PNs, FUN = function(x) {Sys.sleep(0.05);seglengths(x)})
# back to normal
options(nat.progress=NULL)
sl=nlapply(Cell07PNs, FUN = seglengths)
}
}
\seealso{
\code{\link{lapply}}

\code{\link{mapply}}

Other neuronlist: 
\code{\link{*.neuronlist}()},
\code{\link{is.neuronlist}()},
\code{\link{neuronlist}()},
\code{\link{neuronlist-dataframe-methods}},
\code{\link{neuronlistfh}()},
\code{\link{neuronlistz}()},
\code{\link{read.neurons}()},
\code{\link{write.neurons}()}
}
\concept{neuronlist}
