% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nopen3d}
\alias{nopen3d}
\title{Open customised rgl window}
\usage{
nopen3d(bgcol = "white", FOV = 0, ...)
}
\arguments{
  \item{bgcol}{background colour}

  \item{FOV}{field of view}

  \item{...}{additional options passed to open3d}
}
\value{
current rgl device
}
\description{
Pan with right button (Ctrl+click), zoom with middle
(Alt/Meta+click) button. Defaults to a white background and
orthogonal projection (FOV=0)
}
\seealso{
\code{\link{open3d},\link{pan3d}}
}

