% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im3d.R
\name{origin}
\alias{origin}
\title{Return the space origin of a 3D image object}
\usage{
origin(x, ...)
}
\arguments{
\item{x}{Object for which origin should be returned. See 
\code{\link{boundingbox}}.}

\item{\dots}{Additional arguments passed to \code{\link{boundingbox}}}
}
\description{
Defined as the first coordinates (x,y,z) of the bounding box, 
  which in turn matches the nrrd definition of the location of the "centre" 
  of the first voxel.
}
\seealso{
Other im3d: 
\code{\link{as.im3d}()},
\code{\link{boundingbox}()},
\code{\link{im3d-coords}},
\code{\link{im3d-io}},
\code{\link{im3d}()},
\code{\link{imexpand.grid}()},
\code{\link{imslice}()},
\code{\link{is.im3d}()},
\code{\link{mask}()},
\code{\link{projection}()},
\code{\link{threshold}()},
\code{\link{unmask}()},
\code{\link{voxdims}()}
}
\concept{im3d}
