% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{plane_coefficients}
\alias{plane_coefficients}
\title{Find the coefficients of the plane equation}
\usage{
plane_coefficients(p, n)
}
\arguments{
\item{p}{A point on the plane (or N x 3 matrix of multiple points)}

\item{n}{vector normal to the plane (or N x 3 matrix of multiple vectors)}
}
\value{
a matrix with 4 columns \code{a,b,c,d} where \code{ax + by + cz + d =
  0}
}
\description{
Find the coefficients of the plane equation
}
\details{
Both \code{p} and \code{n} can accept multiple points/vectors to 
  calculate many planes at once.
}
\seealso{
Other geometry: \code{\link{intersect_plane}}
}
