% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmtkreg.R
\name{plot3d.cmtkreg}
\alias{plot3d.cmtkreg}
\title{Plot the domain of a CMTK registration}
\usage{
\method{plot3d}{cmtkreg}(x, ..., gridlines = FALSE, plotengine = getOption("nat.plotengine"))
}
\arguments{
\item{x}{A cmtk registration (the path to the registration folder on disk) or
the resulting of reading one in with \code{\link{read.cmtkreg}}.}

\item{...}{Additional arguments passed to \code{\link[rgl]{plot3d}}}

\item{gridlines}{Whether to display gridlines when using plotly as the backend plotting
engine (default: \code{FALSE})}

\item{plotengine}{the plotting backend engine to use either 'rgl' or 'plotly'.}
}
\description{
Plot the domain of a CMTK registration
}
\examples{
\donttest{
testdatadir=system.file("tests/testthat/testdata/cmtk", package="nat")
regpath=file.path(testdatadir,'FCWB_JFRC2_01_warp_level-01.list/')
# only run this if file is present (not always installed)
if(file.exists(regpath)){
plot3d(cmtkreg(regpath))

# or read registration into memory if you want to work with it
reg=read.cmtkreg(regpath)
plot3d(reg)
}
}
}
\seealso{
\code{\link{cmtkreg}}, \code{\link{read.cmtkreg}},
  \code{\link[rgl]{plot3d}}
}
