% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot3d.hxsurf}
\alias{plot3d.hxsurf}
\title{Plot amira surface objects in 3d using rgl}
\usage{
\method{plot3d}{hxsurf}(x, materials = NULL, col = NULL, ...)
}
\arguments{
  \item{x}{An hxsurf surface object}

  \item{materials}{Character vector or \code{\link{regex}}
  naming materials to plot (defaults to all materials in
  x). See \code{\link{subset.hxsurf}}.}

  \item{col}{Character vector specifying colors for the
  materials, or a function that will be called with the
  number of materials to plot. When \code{NULL} (default)
  will use meterial colours defined in Amira (if
  available), or \code{rainbow} otherwise.}

  \item{...}{Additional arguments passed to}
}
\description{
Plot amira surface objects in 3d using rgl
}
\seealso{
\code{\link{read.hxsurf}}

Other hxsurf: \code{\link{as.mesh3d}},
\code{\link{as.mesh3d.hxsurf}}; \code{\link{hxsurf}},
\code{\link{read.hxsurf}}; \code{\link{materials}},
\code{\link{materials.character}},
\code{\link{materials.default}},
\code{\link{materials.hxsurf}};
\code{\link{subset.hxsurf}}; \code{\link{write.hxsurf}}
}

