% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pointsinside}
\alias{pointsinside}
\alias{pointsinside.default}
\title{Find which points of an object are inside a surface}
\usage{
pointsinside(x, surf, ...)

\method{pointsinside}{default}(x, surf, ..., rval = c("logical", "distance",
  "mesh3d"))
}
\arguments{
  \item{x}{An object with 3d points}

  \item{surf}{An \code{hxsurf} or \code{mesh3d} object
  defining the reference surface}

  \item{...}{Additional arguments for methods, eventually
  passed to as.mesh3d}

  \item{rval}{What to return}
}
\value{
A vector of logical values or distances equal to the number
of points in x or the \code{mesh3d} object returned by
\code{Rvcg::vcgClost}.
}
\description{
#FIXME Unfortunately this does not seem to work with the
neuropil surfaces, perhaps because they are not smooth or
because I have misunderstood the definition of signed
distance.
}
\details{
Note that \code{hxsurf} surface objects will be converted
to \code{mesh3d} before being passed to
\code{Rvcg::vcgClost}, so if you are testing repeatedly
against the same surface, it may make sense to pre-convert.
}

