% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngraph.R
\name{prune_in_volume}
\alias{prune_in_volume}
\alias{prune_in_volume.neuron}
\alias{prune_in_volume.neuronlist}
\title{Prune neuron within a volume}
\usage{
prune_in_volume(x, brain, neuropil = NULL, invert = TRUE, ...)

\method{prune_in_volume}{neuron}(x, brain, neuropil = NULL,
  invert = TRUE, ...)

\method{prune_in_volume}{neuronlist}(x, brain, neuropil = NULL,
  invert = TRUE, ...)
}
\arguments{
\item{x}{a \code{neuron} object}

\item{brain}{The \code{\link[nat]{hxsurf}} object containing the neuropil of
interest, e.g. \code{\link[nat.flybrains]{FCWBNP.surf}}. Can also be any other \code{\link[nat]{hxsurf}}
or \code{\link[rgl]{mesh3d}} object, or any object coercible into \code{\link[rgl]{mesh3d}}
by \code{\link[rgl]{as.mesh3d}}}

\item{neuropil}{Character vector specifying the neuropil, which must be queryable from brain$RegionList. 
If NULL (default), then the full object given as \code{brain} will be used for the pruning}

\item{invert}{Logical when \code{TRUE} indicates that points inside the mesh are kept.}

\item{...}{Additional arguments for methods (eventually passed to prune.default)
surface should be pruned.}
}
\description{
Prune neuron within a volume
}
\details{
Prune neuron inside subvolume of a segmented brain object.
}
\examples{
\dontrun{ 
## Interactively shoose which bit of the neuron you wish to keep
### Example requires the package nat.flybrains
LH_arbour = prune_in_volume(x = Cell07PNs, brain = nat.flybrains::IS2NP.surf, neuropil = "LH_L", OmitFailures = TRUE)
} 
}
\seealso{
\code{\link{prune_strahler}}, \code{\link{spine}},
  \code{\link{prune_vertices}}

\code{\link{subset.neuron}}

\code{\link{subset.dotprops}}

Other neuron: \code{\link{neuron}}, \code{\link{ngraph}},
  \code{\link{plot.dotprops}},
  \code{\link{potential_synapses}}, \code{\link{resample}},
  \code{\link{rootpoints}}, \code{\link{spine}},
  \code{\link{subset.neuron}}
}
