% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngraph.R
\name{prune_in_volume}
\alias{prune_in_volume}
\alias{prune_in_volume.neuron}
\alias{prune_in_volume.neuronlist}
\title{Prune neuron within a volume}
\usage{
prune_in_volume(x, brain, neuropil = NULL, invert = TRUE, ...)

\method{prune_in_volume}{neuron}(x, brain, neuropil = NULL,
  invert = TRUE, ...)

\method{prune_in_volume}{neuronlist}(x, brain, neuropil = NULL,
  invert = TRUE, ...)
}
\arguments{
\item{x}{a \code{neuron} object}

\item{brain}{The \code{\link[nat]{hxsurf}} object containing the neuropil of
interest, e.g. \code{\link[nat.flybrains]{FCWBNP.surf}}. Can also be any other \code{\link[nat]{hxsurf}}
or \code{\link[rgl]{mesh3d}} object, or any object coercible into \code{\link[rgl]{mesh3d}}
by \code{\link[rgl]{as.mesh3d}}}

\item{neuropil}{Character vector specifying the neuropil, which must be queryable from brain$RegionList. 
If NULL (default), then the full object given as \code{brain} will be used for the pruning}

\item{invert}{Logical when \code{TRUE} indicates that points inside the mesh are kept.}

\item{...}{Additional arguments for methods (eventually passed to prune.default)
surface should be pruned.}
}
\value{
A pruned neuron/neuronlist object
}
\description{
Prune neuron within a volume
}
\details{
Prune a neuron to be within, or to exclude arbour within, 
a 3D object that can be coecred into the mesh3d data structure
}
\examples{
\dontrun{ 
## Interactively shoose which bit of the neuron you wish to keep
### Example requires the package nat.flybrains
LH_arbour = prune_in_volume(x = Cell07PNs, 
brain = nat.flybrains::IS2NP.surf, 
neuropil = "LH_L", OmitFailures = TRUE)
} 
\dontrun{ 
## Interactively shoose which bit of the neuron you wish to keep
pruned.as.you.like.it = prune_online(Cell07PNs)
}
}
\seealso{
\code{\link{as.neuron.ngraph}}, \code{\link{subset.neuron}}, 
  \code{\link{prune.neuron}}
}
