% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngraph.R
\name{prune_vertices}
\alias{prune_vertices}
\title{Prune selected vertices from a neuron}
\usage{
prune_vertices(x, verticestoprune)
}
\arguments{
\item{x}{A neuron to prune}

\item{verticestoprune}{An integer vector describing which vertices to remove.
The special signalling value of \code{NA} drops all vertices with invalid X
locations.}
}
\description{
Prune selected vertices from a neuron
}
\details{
uses the \code{ngraph} representation of the neuron to remove
}

