% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.cmtkreg}
\alias{read.cmtkreg}
\title{Read a CMTK format registration}
\usage{
read.cmtkreg(filename, ReturnRegistrationOnly = FALSE, ...)
}
\arguments{
  \item{filename}{Path to a CMTK registration file}

  \item{ReturnRegistrationOnly}{When FALSE (default) will
  not attempt to extract the registration element from the
  registration file.}

  \item{...}{Additional arguments passed to read.cmtk}
}
\description{
Read a CMTK format registration
}
\seealso{
Other cmtk-io: \code{\link{cmtklandmarks}},
\code{\link{read.cmtklandmarks}},
\code{\link{write.cmtklandmarks}}; \code{\link{read.cmtk}};
\code{\link{write.cmtkreg}}; \code{\link{write.cmtk}}
}

