% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron-io-fiji.R
\name{read.neuron.fiji}
\alias{read.neuron.fiji}
\title{Read a neuron saved by Fiji's Simple Neurite Tracer Plugin}
\usage{
read.neuron.fiji(f, ..., simplify = TRUE, Verbose = FALSE)
}
\arguments{
\item{f}{Path to a file}

\item{...}{Additional arguments passed to \code{\link[XML]{xmlParse}}.}

\item{simplify}{Whether to return a single neuron as a \code{neuron} object 
rather than a \code{neuronlist} of length 1.}

\item{Verbose}{Whether to print status messages during parsing.}
}
\description{
Read a neuron saved by Fiji's Simple Neurite Tracer Plugin
}
\details{
This is an XML based format so parsing it depends on installation of
  the suggested XML package.
}
\references{
\url{http://fiji.sc/Simple_Neurite_Tracer} 
  \url{http://fiji.sc/Simple_Neurite_Tracer:_.traces_File_Format}
}

