% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuron-io-neuroml.R
\name{read.neuron.neuroml}
\alias{read.neuron.neuroml}
\title{Read one or more neurons from a NeuroML v1 file}
\usage{
read.neuron.neuroml(f, ..., AlwaysReturnNeuronList = FALSE)
}
\arguments{
\item{f}{Path to a NeuroML format XML file}

\item{...}{Additional arguments passed to read.morphml (and on to
\code{\link[XML]{xmlParse}})}

\item{AlwaysReturnNeuronList}{See \bold{Value} section (default \code{FALSE})}
}
\value{
When the XML file contains only 1 cell \emph{and}
  \code{AlwaysReturnNeuronList=FALSE}, a \code{\link{neuron}} object,
  otherwise a \code{\link{neuronlist}} containing one or more neurons.
}
\description{
Read one or more neurons from a NeuroML v1 file
}
\references{
\url{http://www.neuroml.org/specifications}
}
\seealso{
\code{\link{read.morphml}}
}

