\name{read.neuronlistfh}
\alias{read.neuronlistfh}
\title{Read a local, or remote, neuronlistfh object saved to a file.}
\usage{
read.neuronlistfh(file, localdir = NULL)
}
\arguments{
  \item{file}{The file path of the neuronlistfh object. Can
  be local, or remote (via http or ftp).}

  \item{localdir}{If the file is to be fetched from a
  remote location, this is the folder in which downloaded
  objects will be stored.}
}
\description{
Read a local, or remote, neuronlistfh object saved to a
file.
}

