% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuron.R
\name{resample}
\alias{resample}
\alias{resample.neuron}
\title{Resample an object with a new spacing}
\usage{
resample(x, ...)

\method{resample}{neuron}(x, stepsize, ...)
}
\arguments{
\item{x}{An object to resample}

\item{...}{Additional arguments passed to methods}

\item{stepsize}{The new spacing along the tracing}
}
\description{
Resample an object with a new spacing

resample a neuron with a new spacing
}
\details{
\code{resample.neuron} Floating point columns including X,Y,Z,W will
  be interpolated using linear interpolation, while integer or factor columns
  will be interpolated using constant interpolation. See \code{\link{approx}}
  for details.
}
\seealso{
\code{\link{approx}}, \code{\link{seglengths}}
}

