% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-nodes.R
\name{rootpoints}
\alias{rootpoints}
\alias{rootpoints.default}
\alias{rootpoints.neuron}
\alias{rootpoints.igraph}
\alias{branchpoints}
\alias{branchpoints.default}
\alias{branchpoints.neuron}
\alias{branchpoints.igraph}
\alias{endpoints}
\alias{endpoints.neuron}
\alias{endpoints.igraph}
\alias{endpoints.default}
\title{Return the root, branch, or end points of a neuron or graph}
\usage{
rootpoints(x, ...)

\method{rootpoints}{default}(x, ...)

\method{rootpoints}{neuron}(x, subtrees = 1, ...)

\method{rootpoints}{igraph}(x, ...)

branchpoints(x, ...)

\method{branchpoints}{default}(x, ...)

\method{branchpoints}{neuron}(x, subtrees = 1, ...)

\method{branchpoints}{igraph}(x, ...)

endpoints(x, ...)

\method{endpoints}{neuron}(x, subtrees = 1, ...)

\method{endpoints}{igraph}(x, ...)

\method{endpoints}{default}(x, ...)
}
\arguments{
\item{x}{Neuron or other object (e.g. \code{igraph}) which might have roots}

\item{...}{Further arguments passed to methods (for \code{\link{ngraph}} or
\code{igraph} objects eventually \code{\link{graph.nodes}})).}

\item{subtrees}{Integer index of the fully connected subtree in 
\code{x$SubTrees}. Only applicable when a \code{neuron} consists of 
multiple unconnected subtrees.}
}
\value{
FIXME Raw indices (in range 1:N) of vertices when \code{x} is a
  neuron, integer point identifier (aka PointNo) otherwise.
}
\description{
\code{rootpoints} returns the root point(s) (one per tree, often
  the soma).

\code{branchpoints} returns the branch points.

\code{endpoints} returns the end points (aka leaf nodes); the
  root point will be returned if it also a leaf node.
}
\details{
A neuron may have multiple subtrees and therefore multiple roots. At
  present there is discrepancy between the \code{*.neuron} and
  \code{*.igraph} methods. For \code{neuron}s we return the node indices, for
  \code{igraph}/\code{ngraph} objects the node identifiers (aka
  names/PointNo)

\code{branchpoints.neuron} returns a list if more than one subtree
  is specified
}
\examples{
rootpoints(Cell07PNs[[1]])
endpoints(Cell07PNs[[1]])
}
\seealso{
\code{\link{graph.nodes}}, \code{\link{ngraph}}

Other neuron: 
\code{\link{neuron}()},
\code{\link{ngraph}()},
\code{\link{plot.dotprops}()},
\code{\link{potential_synapses}()},
\code{\link{prune}()},
\code{\link{resample}()},
\code{\link{spine}()},
\code{\link{subset.neuron}()}
}
\concept{neuron}
