% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{select_points}
\alias{select_points}
\title{Interactively select 3D points in space}
\usage{
select_points(points, plot3d = NULL)
}
\arguments{
\item{points}{a matrix of 3D points to plot}

\item{plot3d}{additional object that can be plotted using \code{rgl::plot3d}, to plot alongside points (e.g. for context)}
}
\value{
A matrix describing selected 3D points
}
\description{
Plot a set of 3D points in space and select a subet of them interactively, using an rgl window
}
\examples{
\dontrun{ 
# Make a model based off of fly olfactory projection neuron arbours
selected_points = select_points(xyzmatrix(Cell07PNs))
} 
}
\seealso{
\code{\link{prune_online}}
}
