% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{select_points}
\alias{select_points}
\title{Select 3D points in space}
\usage{
select_points(points, plot3d = NULL)
}
\arguments{
\item{points}{a mxn matrix}

\item{plot3d}{an object to pass to the rgl::plot3d function, for context}
}
\value{
A mxn matrix
}
\description{
Select and deselect 3D points in space interactively. Black = selected, red = deselected
}
\examples{
\dontrun{ 
# Make a model based off of fly olfactory projection neuron arbours
chosen = select_points(xyzmatrix(Cell07PNs))
} 
}
