% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron.R
\name{stitch_neuron}
\alias{stitch_neuron}
\title{Stitch two neurons together at their closest endpoint}
\usage{
stitch_neuron(a, b)
}
\arguments{
\item{a, b}{Neurons to join together}
}
\description{
Stitch two neurons together at their closest endpoint
}
\details{
This function joins two neurons at their nearest point (only one).
  Let's say you have two neurons a and b. a and b will be joined at one point that are closest to each other.
  However, when say there are multiple points at a and b which are closer and could be joined, 
  then do not use this function, use the function \code{\link{stitch_neurons_mst}}, 
  which is slower but will merge at multiple points.
  Note that for CATMAID neurons the neuron with the soma tag will be
  treated as the first (master neuron). Furthermore in this case the PointNo
  (aka node id) should already be unique. Otherwise it will be adjusted to
  ensure this.
}
\examples{
\dontrun{
library(catmaid)
dl1=read.neuron.catmaid(catmaid_skids('annotation:DL1')[1])
dl1_main=simplify_neuron(dl1, n = 1, invert = F)
dl1_branches=simplify_neuron(dl1, n = 1, invert = T)
dl1_whole = stitch_neuron(dl1_main,dl1_branches)

}
}
\seealso{
\code{\link{stitch_neurons}}
}
\author{
Gregory Jefferis \email{jefferis@gmail.com}
}
