% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron.R
\name{stitch_neurons_mst}
\alias{stitch_neurons_mst}
\title{Stitch multiple fragments into single neuron using minimum spanning tree}
\usage{
stitch_neurons_mst(x, thresh_el = 1000, k = 10L)
}
\arguments{
\item{x}{Fragments that could be \code{\link{neuronlist}} or a single neuron
with multiple unconnected subtrees.}

\item{thresh_el}{The threshold distance (units in microns) above which new vertices 
will not be connected (default=1000, set to NULL to disable this feature). This parameter prevents the 
merging of vertices that are so far away from the main neuron such that they are likely to be spurious.}

\item{k}{The number of nearest neighbours to consider when trying to merge different clusters.}
}
\value{
A single \code{neuron} object containing all input fragments.
}
\description{
Stitch multiple fragments into single neuron using minimum spanning tree
}
\details{
The neurons are joined using the minimum spanning tree i.e. the tree
  that minimises the sum of edge weights (here, the Euclidean distance). The
  neuron is rooted in the largest cluster; if this cluster contained the
  original root of the neuron, then this should be retained.
}
\examples{

n=Cell07PNs[['ECA34L']]
n_main=simplify_neuron(n, n = 10)
n_branches=simplify_neuron(n, n = 10, invert = TRUE)

# this is the setup
plot(n_main, col='red', WithNodes=FALSE)
plot(n_branches, col='blue', add=TRUE, WithNodes=FALSE)

# make a neuronlist containing the two fragments
nl=neuronlist(n_main, n_branches)
# and stitch those
n_stitched=stitch_neurons_mst(nl)

\dontrun{
# look at the neurons in 3D - they appear identical in this case
plot3d(n, alpha=.5, col='cyan', WithNodes=FALSE)
plot3d(n_stitched, alpha=.5, col='red', WithNodes=FALSE)
}
}
\seealso{
\code{\link{simplify_neuron}}, \code{\link{spine}},
  \code{\link{ngraph}}, \code{igraph::\link[igraph]{mst}}
}
\author{
Sridhar Jagannathan \email{j.sridharrajan@gmail.com}
}
