\name{subset.hxsurf}
\alias{subset.hxsurf}
\title{Subset hxsurf object to specified regions}
\usage{
\method{subset}{hxsurf}(x, subset, ...)
}
\arguments{
  \item{x}{A dotprops object}

  \item{subset}{Character vector specifying regions to
  keep}

  \item{...}{Additional parameters (currently ignored)}
}
\value{
subsetted hxsurf object
}
\description{
Subset hxsurf object to specified regions
}

