% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subset.hxsurf}
\alias{subset.hxsurf}
\title{Subset hxsurf object to specified regions}
\usage{
\method{subset}{hxsurf}(x, subset = NULL, drop = FALSE, ...)
}
\arguments{
  \item{x}{A dotprops object}

  \item{subset}{Character vector specifying regions to
  keep}

  \item{drop}{Whether to drop unused vertices after
  subsetting}

  \item{...}{Additional parameters (currently ignored)}
}
\value{
subsetted hxsurf object
}
\description{
Subset hxsurf object to specified regions
}
\seealso{
Other hxsurf: \code{\link{as.mesh3d}},
\code{\link{as.mesh3d.hxsurf}}; \code{\link{hxsurf}},
\code{\link{read.hxsurf}}; \code{\link{plot3d.hxsurf}};
\code{\link{write.hxsurf}}
}

