% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuron.R
\name{subset.neuron}
\alias{subset.neuron}
\title{Subset neuron by keeping only vertices that match given conditions}
\usage{
\method{subset}{neuron}(x, subset, ...)
}
\arguments{
\item{x}{A neuron object}

\item{subset}{A subset of points defined by indices, an expression, or a
function (see Details)}

\item{...}{Additional parameters (passsed on to \code{\link{prune_vertices}})}
}
\value{
subsetted neuron
}
\description{
Subset neuron by keeping only vertices that match given conditions
}
\details{
\code{subset} defines which vertices of the neuron to keep and is
  one of \itemize{

  \item logical or numeric indices, in which case these are simply used to
  index the vertices in the order of the

  \item a function (which is called with the 3d points array and returns T/F
  vector)

  \item an expression evaluated in the context of the \code{x$d} data.frame
  containing the SWC specification of the points and connectivity of the
  neuron. This can therefore refer e.g. to the x,y,z location of vertices in
  the neuron.

  }
}
\examples{
n=Cell07PNs[[1]]
# keep vertices if their X location is > 2000
n1=subset(n, X>200)
# diameter of neurite >1
n2=subset(n, W>1)
# first 50 nodes
n3=subset(n,1:50)

\dontrun{
plot3d(n)
s3d=select3d()
n4=subset(n, s3d(xyzmatrix(n)))
# special case of previous version
n5=subset(n, s3d)
stopifnot(all.equal(n4,n5))
# keep the points that were removed from n1
n4.not=subset(n,Negate(s3d))
# vertices with x position > 100 and inside the selector function
n6=subset(n,X>100 & s3d(X,Y,Z))

## subset each neuron object in a whole neuronlist
n10=Cell07PNs[1:10]
plot3d(n10, lwd=0.5, col='grey')
n10.crop = nlapply(n10, subset, X>250)
plot3d(n10.crop, col='red')
}
}
\seealso{
\code{\link{prune.neuron}}, \code{\link{prune_vertices}},
  \code{\link{subset.dotprops}}
}

