% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write.cmtk}
\alias{write.cmtk}
\title{Write a suitable list to a CMTK TypedStream file on disk}
\usage{
write.cmtk(l, con, gzip = FALSE, version = NA_character_)
}
\arguments{
  \item{l}{Appropriately formatted list}

  \item{con}{A character string specifying a path or a
  connection}

  \item{gzip}{Whether to gzip output file (default FALSE)}

  \item{version}{TYPEDSTREAM version number, defaults to
  \code{"1.1"} if not specified in the version attribute of
  \code{l}.}
}
\description{
This is probaly only of interest to developers. End users
will probably wish to use more specific functions such as
write.cmtkreg for writing out registrations.
}
\details{
NB a version specified on the command line overrides one
encoded as an attribute in the input list.
}
\seealso{
Other cmtk-io: \code{\link{cmtklandmarks}},
\code{\link{read.cmtklandmarks}},
\code{\link{write.cmtklandmarks}};
\code{\link{read.cmtkreg}}; \code{\link{read.cmtk}};
\code{\link{write.cmtkreg}}
}

